@@ -2,60 +2,111 @@
 
 import functools
 import json
+import uuid
 import warnings
 from collections import OrderedDict
 from io import StringIO
-from typing import List, Union, Dict, Iterable
+from typing import List, Union, Dict, Iterable, Tuple
 
-import pandas as pd
-from requests import Response
-
+from TM1py.Exceptions.Exceptions import TM1pyException
+from TM1py.Objects.MDXView import MDXView
+from TM1py.Objects.Process import Process
 from TM1py.Services.ObjectService import ObjectService
 from TM1py.Services.RestService import RestService
-from TM1py.Utils import Utils, CaseAndSpaceInsensitiveSet, format_url
+from TM1py.Services.ViewService import ViewService
+from TM1py.Utils import Utils, CaseAndSpaceInsensitiveSet, format_url, add_url_parameters
 from TM1py.Utils.Utils import build_pandas_dataframe_from_cellset, dimension_name_from_element_unique_name, \
-    CaseAndSpaceInsensitiveTuplesDict
+    CaseAndSpaceInsensitiveDict, wrap_in_curly_braces, CaseAndSpaceInsensitiveTuplesDict, abbreviate_mdx, \
+    build_csv_from_cellset_dict, require_version, require_pandas, build_cellset_from_pandas_dataframe, \
+    case_and_space_insensitive_equals, get_cube, resembles_mdx, require_admin
+from mdxpy import MdxHierarchySet, MdxBuilder, Member
+from requests import Response
+
+try:
+    import pandas as pd
+
+    _has_pandas = True
+except ImportError:
+    _has_pandas = False
 
 
 def tidy_cellset(func):
-    """ Higher Order Function to tidy up cellset after usage
+    """ Higher order function to tidy up cellset after usage
     """
 
     @functools.wraps(func)
     def wrapper(self, cellset_id, *args, **kwargs):
         try:
             return func(self, cellset_id, *args, **kwargs)
+
         finally:
             if kwargs.get("delete_cellset", True):
-                self.delete_cellset(cellset_id=cellset_id)
+                sandbox_name = kwargs.get("sandbox_name", None)
+                if sandbox_name is not None:
+                    self.delete_cellset(cellset_id=cellset_id, sandbox_name=sandbox_name)
+                else:
+                    self.delete_cellset(cellset_id=cellset_id)
+
+    return wrapper
+
+
+def manage_transaction_log(func):
+    """ Control state of transaction log during and after write operation for a given cube through:
+    `deactivate_transaction_log` and `reactivate_transaction_log`.
+
+    Decorated function must have either `cube_name` or `mdx` as first argument or keyword argument
+    """
+
+    @functools.wraps(func)
+    def wrapper(self, *args, **kwargs):
+        if "cube_name" in kwargs:
+            cube_name = kwargs["cube_name"]
+        elif "mdx" in kwargs:
+            cube_name = get_cube(kwargs["mdx"])
+        else:
+            arg = args[0]
+            if resembles_mdx(arg):
+                cube_name = get_cube(arg)
+            else:
+                cube_name = arg
+
+        try:
+            if kwargs.get("deactivate_transaction_log", False):
+                self.deactivate_transactionlog(cube_name)
+            return func(self, *args, **kwargs)
+
+        finally:
+            if kwargs.get("reactivate_transaction_log", False):
+                self.activate_transactionlog(cube_name)
 
     return wrapper
 
 
 class CellService(ObjectService):
     """ Service to handle Read and Write operations to TM1 cubes
-    
+
     """
 
     def __init__(self, tm1_rest: RestService):
         """
-        
+
         :param tm1_rest: instance of RestService
         """
         super().__init__(tm1_rest)
 
-    def get_value(self, cube_name: str, element_string: str, dimensions: List[str] = None,
+    def get_value(self, cube_name: str, element_string: str, dimensions: List[str] = None, sandbox_name: str = None,
                   **kwargs) -> Union[str, float]:
         """ Element_String describes the Dimension-Hierarchy-Element arrangement
-            
+
         :param cube_name: Name of the cube
         :param element_string: "Hierarchy1::Element1 && Hierarchy2::Element4, Element9, Element2"
             - Dimensions are not specified! They are derived from the position.
-            - The , seperates the element-selections
+            - The , separates the element-selections
             - If more than one hierarchy is selected per dimension && splits the elementselections
             - If no Hierarchy is specified. Default Hierarchy will be addressed
         :param dimensions: List of dimension names in correct order
-        :return: 
+        :param sandbox_name: str
+        :return:
         """
         mdx_template = "SELECT {} ON ROWS, {} ON COLUMNS FROM [{}]"
         mdx_rows_list = []
@@ -64,7 +115,7 @@ class CellService(ObjectService):
             dimensions = CubeService(self._rest).get(cube_name).dimensions
         element_selections = element_string.split(',')
         # Build the ON ROWS statement:
-        # Loop through the comma seperated element selection, except for the last one
+        # Loop through the comma separated element selection, except for the last one
         for dimension_name, element_selection in zip(dimensions[:-1], element_selections[:-1]):
             if "&&" not in element_selection:
                 mdx_rows_list.append("{[" + dimension_name + "].[" + dimension_name + "].[" + element_selection + "]}")
@@ -86,7 +137,7 @@ class CellService(ObjectService):
         # Construct final MDX
         mdx = mdx_template.format(mdx_rows, mdx_columns, cube_name)
         # Execute MDX
-        cellset = dict(self.execute_mdx(mdx, **kwargs))
+        cellset = dict(self.execute_mdx(mdx=mdx, sandbox_name=sandbox_name, **kwargs))
         return next(iter(cellset.values()))["Value"]
 
     def relative_proportional_spread(
@@ -96,6 +147,7 @@ class CellService(ObjectService):
             unique_element_names: Iterable[str],
             reference_unique_element_names: Iterable[str],
             reference_cube: str = None,
+            sandbox_name: str = None,
             **kwargs) -> Response:
         """ Execute relative proportional spread
 
@@ -104,6 +156,7 @@ class CellService(ObjectService):
         :param unique_element_names: target cell coordinates as unique element names (e.g. ["[d1].[c1]","[d2].[e3]"])
         :param reference_cube: name of the reference cube. Can be None
         :param reference_unique_element_names: reference cell coordinates as unique element names
+        :param sandbox_name: str
         :return:
         """
         mdx = """
@@ -111,7 +164,7 @@ class CellService(ObjectService):
         {{ {rows} }} ON 0
         FROM [{cube}]
         """.format(rows="}*{".join(unique_element_names), cube=cube)
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
 
         payload = {
             "BeginOrdinal": 0,
@@ -125,16 +178,19 @@ class CellService(ObjectService):
                     "Dimensions('{}')/Hierarchies('{}')/Elements('{}')",
                     *Utils.dimension_hierarchy_element_tuple_from_unique_name(unique_element_name)))
 
-        return self._post_against_cellset(cellset_id=cellset_id, payload=payload, delete_cellset=True, **kwargs)
+        return self._post_against_cellset(cellset_id=cellset_id, payload=payload, delete_cellset=True,
+                                          sandbox_name=sandbox_name, **kwargs)
 
     def clear_spread(
             self,
             cube: str,
             unique_element_names: Iterable[str],
+            sandbox_name: str = None,
             **kwargs) -> Response:
         """ Execute clear spread
         :param cube: name of the cube
         :param unique_element_names: target cell coordinates as unique element names (e.g. ["[d1].[c1]","[d2].[e3]"])
+        :param sandbox_name: str
         :return:
         """
         mdx = """
@@ -142,7 +198,7 @@ class CellService(ObjectService):
         {{ {rows} }} ON 0
         FROM [{cube}]
         """.format(rows="}*{".join(unique_element_names), cube=cube)
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
 
         payload = {
             "BeginOrdinal": 0,
@@ -154,18 +210,82 @@ class CellService(ObjectService):
                     "Dimensions('{}')/Hierarchies('{}')/Elements('{}')",
                     *Utils.dimension_hierarchy_element_tuple_from_unique_name(unique_element_name)))
 
-        return self._post_against_cellset(cellset_id=cellset_id, payload=payload, delete_cellset=True, **kwargs)
+        return self._post_against_cellset(cellset_id=cellset_id, payload=payload, delete_cellset=True,
+                                          sandbox_name=sandbox_name, **kwargs)
+
+    @require_admin
+    @require_version(version="11.7")
+    def clear(self, cube: str, **kwargs):
+        """
+        Takes the cube name and keyword argument pairs of dimensions and expressions:
+        `tm1.cells.clear(cube="Sales", product="{[Product].[ABC]}", time="{[Time].[2020].Children}")`
+
+        :param cube: name of the cube
+        :param kwargs: keyword argument pairs of dimension names and mdx set expressions
+        :return:
+        """
+        cube_service = self.get_cube_service()
+        dimension_names = CaseAndSpaceInsensitiveSet(*cube_service.get_dimension_names(cube_name=cube))
+        dimension_expression_pairs = CaseAndSpaceInsensitiveDict()
+
+        for kwarg in kwargs:
+            if kwarg in dimension_names:
+                dimension_expression_pairs[kwarg] = wrap_in_curly_braces(kwargs[kwarg])
+
+        for dimension_name in dimension_names:
+            if dimension_name not in dimension_expression_pairs:
+                expression = MdxHierarchySet.tm1_subset_all(dimension_name).filter_by_level(0).to_mdx()
+                dimension_expression_pairs[dimension_name] = expression
+
+        mdx_builder = MdxBuilder.from_cube(cube).columns_non_empty()
+        for dimension, expression in dimension_expression_pairs.items():
+            hierarchy_set = MdxHierarchySet.from_str(dimension=dimension, hierarchy=dimension, mdx=expression)
+            mdx_builder.add_hierarchy_set_to_column_axis(hierarchy_set)
+
+        return self.clear_with_mdx(cube=cube, mdx=mdx_builder.to_mdx(), **kwargs)
+
+    @require_admin
+    @require_version(version="11.7")
+    def clear_with_mdx(self, cube: str, mdx: str, **kwargs):
+        """ clear a slice in a cube based on an MDX query.
+        Function requires admin permissions, since TM1py uses an unbound TI with a `ViewZeroOut` statement.
+
+        :param cube: name of the cube
+        :param mdx: a valid MDX query
+        :param kwargs:
+        :return:
+        """
+        from TM1py import ProcessService
+        process_service = ProcessService(self._rest)
+        view_service = ViewService(self._rest)
+
+        view_name = "".join(['}TM1py', str(uuid.uuid4())])
+        view_service.create(MDXView(cube_name=cube, view_name=view_name, MDX=mdx))
+
+        try:
+            code = f"ViewZeroOut('{cube}','{view_name}');"
+            process = Process(name="")
+            process.prolog_procedure = code
+
+            success, _, _ = self.execute_unbound_process(process, **kwargs)
+            if not success:
+                raise TM1pyException(f"Failed to clear cube: '{cube}' with mdx: '{abbreviate_mdx(mdx, 100)}'")
+        finally:
+            if view_service.exists(cube, view_name, private=False):
+                view_service.delete(cube, view_name, private=False)
 
     @tidy_cellset
-    def _post_against_cellset(self, cellset_id: str, payload: Dict, **kwargs) -> Response:
+    def _post_against_cellset(self, cellset_id: str, payload: Dict, sandbox_name: str = None, **kwargs) -> Response:
         """ Execute a post request against a cellset
 
         :param cellset_id:
         :param payload:
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
         url = format_url("/api/v1/Cellsets('{}')/tm1.Update", cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         return self._rest.POST(url=url, data=json.dumps(payload), **kwargs)
 
     def get_dimension_names_for_writing(self, cube_name: str, **kwargs) -> List[str]:
@@ -180,19 +300,51 @@ class CellService(ObjectService):
         dimensions = cube_service.get_dimension_names(cube_name, True, **kwargs)
         return dimensions
 
+    @require_pandas
+    def write_dataframe(self, cube_name: str, data: 'pd.DataFrame', dimensions: Iterable[str] = None,
+                        sandbox_name: str = None,
+                        **kwargs) -> Response:
+        """
+        Function expects same shape as `execute_mdx_dataframe` returns.
+        Column order must match dimensions in the target cube with an additional column for the values.
+        Column names are not relevant.
+
+        :param cube_name:
+        :param data:
+        :param dimensions:
+        :param sandbox_name: str
+        :param kwargs:
+        :return:
+        """
+        if not isinstance(data, pd.DataFrame):
+            raise ValueError("argument 'data' must of type DataFrame")
+
+        if not dimensions:
+            dimensions = self.get_dimension_names_for_writing(cube_name=cube_name)
+
+        if not len(data.columns) == len(dimensions) + 1:
+            raise ValueError("Number of columns in 'data' DataFrame must be number of dimensions in cube + 1")
+
+        cells = build_cellset_from_pandas_dataframe(data)
+
+        return self.write_values(cube_name=cube_name, cellset_as_dict=cells, dimensions=dimensions,
+                                 sandbox_name=sandbox_name, **kwargs)
+
     def write_value(self, value: Union[str, float], cube_name: str, element_tuple: Iterable,
-                    dimensions: Iterable[str] = None, **kwargs) -> Response:
+                    dimensions: Iterable[str] = None, sandbox_name: str = None, **kwargs) -> Response:
         """ Write value into cube at specified coordinates
 
         :param value: the actual value
         :param cube_name: name of the target cube
         :param element_tuple: target coordinates
         :param dimensions: optional. Dimension names in their natural order. Will speed up the execution!
+        :param sandbox_name: str
         :return: response
         """
         if not dimensions:
             dimensions = self.get_dimension_names_for_writing(cube_name=cube_name)
         url = format_url("/api/v1/Cubes('{}')/tm1.Update", cube_name)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         body_as_dict = OrderedDict()
         body_as_dict["Cells"] = [{}]
         body_as_dict["Cells"][0]["Tuple@odata.bind"] = [
@@ -203,19 +355,154 @@ class CellService(ObjectService):
         data = json.dumps(body_as_dict, ensure_ascii=False)
         return self._rest.POST(url=url, data=data, **kwargs)
 
-    def write_values(self, cube_name: str, cellset_as_dict: Dict, dimensions: Iterable[str] = None,
-                     **kwargs) -> Response:
-        """ Write values in cube.  
-        For cellsets with > 1000 cells look into "write_values_through_cellset"
+    def write(self, cube_name: str, cellset_as_dict: Dict, dimensions: Iterable[str] = None, increment: bool = False,
+              deactivate_transaction_log: bool = False, reactivate_transaction_log: bool = False,
+              sandbox_name: str = None, use_ti=False, **kwargs):
+        """ Write values to a cube
+
+        Same signature as `write_values` method, but faster since it uses `write_values_through_cellset`
+        behind the scenes.
+
+        Supports incrementing cell values through optional `increment` argument
+        Spreading through spreading shortcuts is not supported!
 
         :param cube_name: name of the cube
         :param cellset_as_dict: {(elem_a, elem_b, elem_c): 243, (elem_d, elem_e, elem_f) : 109}
         :param dimensions: optional. Dimension names in their natural order. Will speed up the execution!
+        :param increment: increment or update cell values
+        :param deactivate_transaction_log: deactivate before writing
+        :param reactivate_transaction_log: reactivate after writing
+        :param sandbox_name: str
+        :param use_ti: Use unbound process to write. Requires admin permissions. causes massive performance improvement.
+        :return:
+        """
+        if use_ti:
+            return self.write_through_unbound_process(
+                cube_name=cube_name,
+                cellset_as_dict=cellset_as_dict,
+                increment=increment,
+                sandbox_name=sandbox_name,
+                deactivate_transaction_log=deactivate_transaction_log,
+                reactivate_transaction_log=reactivate_transaction_log,
+                **kwargs)
+
+        if not dimensions:
+            dimensions = self.get_dimension_names_for_writing(cube_name=cube_name, **kwargs)
+
+        values = []
+        query = MdxBuilder.from_cube(cube_name)
+        for coordinates, value in cellset_as_dict.items():
+            members = (Member.of(dimension, element) for dimension, element in zip(dimensions, coordinates))
+            query.add_member_tuple_to_columns(*members)
+            values.append(value)
+        mdx = query.to_mdx()
+
+        self.write_values_through_cellset(
+            mdx=mdx,
+            values=values,
+            increment=increment,
+            deactivate_transaction_log=deactivate_transaction_log,
+            reactivate_transaction_log=reactivate_transaction_log,
+            sandbox_name=sandbox_name,
+            **kwargs)
+
+    @require_admin
+    @manage_transaction_log
+    def write_through_unbound_process(self, cube_name: str, cellset_as_dict: Dict, increment: bool = False,
+                                      sandbox_name: str = None, precision=8, **kwargs):
+        if sandbox_name:
+            raise NotImplementedError("Function does not support writing to sandboxes yet")
+
+        successes = list()
+        statements = list()
+
+        cube_service = self.get_cube_service()
+        element_service = self.get_element_service()
+
+        measure_dimension = cube_service.get_measure_dimension(cube_name=cube_name)
+
+        measure_dimension_elements = element_service.get_element_types(
+            dimension_name=measure_dimension,
+            hierarchy_name=measure_dimension,
+            skip_consolidations=False)
+
+        for coordinates, value in cellset_as_dict.items():
+            # use default 'Numeric' so that not existing elements trigger minor error during TI execution
+            element_type = measure_dimension_elements.get(coordinates[-1], 'Numeric')
+
+            if element_type == 'String':
+                function_str = 'CellPutS('
+                value_str = f"'{value}'"
+
+            # by default assume numeric, to trigger minor errors on write operations to C elements
+            else:
+                function_str = "CellIncrementN(" if increment else "CellPutN("
+                # number strings must not exceed float range
+                if isinstance(value, str):
+                    try:
+                        value_str = format(float(value), f'.{precision}f')
+                    except ValueError:
+                        value_str = f'{value}'
+                else:
+                    value_str = format(value, f'.{precision}f')
+
+            statement = "".join([
+                function_str,
+                value_str,
+                f",'{cube_name}',",
+                ",".join(f"'{element}'" for element in coordinates),
+                ");"])
+            statements.append(statement)
+
+        chunk = list()
+        for n, statement in enumerate(statements):
+            chunk.append(statement)
+            if n > 0 and n % Process.MAX_STATEMENTS == 0:
+                process = Process(name="", prolog_procedure="\r".join(chunk))
+                success, _, _ = self.execute_unbound_process(process, **kwargs)
+                successes.append(success)
+                chunk = list()
+
+        process = Process(name="", prolog_procedure="\r".join(chunk))
+        success, _, _ = self.execute_unbound_process(process, **kwargs)
+        successes.append(success)
+
+        if not all(successes):
+            raise TM1pyException(f"{len(successes) - sum(successes)} out of {len(successes)} unbound processes failed")
+
+    def get_element_service(self):
+        from TM1py import ElementService
+        return ElementService(self._rest)
+
+    def get_cube_service(self):
+        from TM1py import CubeService
+        return CubeService(self._rest)
+
+    def execute_unbound_process(self, process: Process, **kwargs) -> Tuple[bool, str, str]:
+        from TM1py import ProcessService
+        process_service = ProcessService(self._rest)
+
+        return process_service.execute_process_with_return(process, **kwargs)
+
+    @manage_transaction_log
+    def write_values(self, cube_name: str, cellset_as_dict: Dict, dimensions: Iterable[str] = None,
+                     sandbox_name: str = None, **kwargs) -> Response:
+        """ Write values to a cube
+
+        For cellsets with > 1000 cells look into `write` or `write_values_through_cellset`
+        Supports spreading shortcuts
+
+        :param cube_name: name of the cube
+        :param cellset_as_dict: {(elem_a, elem_b, elem_c): 243, (elem_d, elem_e, elem_f) : 109}
+        :param dimensions: optional. Dimension names in their natural order. Will speed up the execution!
+        :param sandbox_name: str
         :return: Response
         """
         if not dimensions:
-            dimensions = self.get_dimension_names_for_writing(cube_name=cube_name)
+            dimensions = self.get_dimension_names_for_writing(cube_name=cube_name, **kwargs)
         url = format_url("/api/v1/Cubes('{}')/tm1.Update", cube_name)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
+
         updates = []
         for element_tuple, value in cellset_as_dict.items():
             body_as_dict = OrderedDict()
@@ -231,95 +518,133 @@ class CellService(ObjectService):
         updates = '[' + ','.join(updates) + ']'
         return self._rest.POST(url=url, data=updates, **kwargs)
 
-    def write_values_through_cellset(self, mdx: str, values: List, **kwargs) -> Response:
+    @manage_transaction_log
+    def write_values_through_cellset(self, mdx: str, values: Iterable, increment: bool = False,
+                                     sandbox_name: str = None, **kwargs) -> Response:
         """ Significantly faster than write_values function
+
         Cellset gets created according to MDX Expression. For instance:
-        [[61, 29 ,13], 
-        [42, 54, 15], 
+        [[61, 29 ,13],
+        [42, 54, 15],
         [17, 28, 81]]
-        
-        Each value in the cellset can be addressed through its position: The ordinal integer value. 
+
+        Each value in the cellset can be addressed through its position: The ordinal integer value.
         Ordinal-enumeration goes from top to bottom from left to right
         Number 61 has Ordinal 0, 29 has Ordinal 1, etc.
 
-        The order of the iterable determines the insertion point in the cellset. 
+        The order of the iterable determines the insertion point in the cellset.
         For instance:
         [91, 85, 72, 68, 51, 42, 35, 28, 11]
 
         would lead to:
-        [[91, 85 ,72], 
-        [68, 51, 42], 
+        [[91, 85 ,72],
+        [68, 51, 42],
         [35, 28, 11]]
 
         When writing large datasets into TM1 Cubes it can be convenient to call this function asynchronously.
-        
+
         :param mdx: Valid MDX Expression.
         :param values: List of values. The Order of the List/ Iterable determines the insertion point in the cellset.
-        :return: 
+        :param increment: increment or update cells
+        :param sandbox_name: str
+        :return:
         """
-        cellset_id = self.create_cellset(mdx, **kwargs)
-        return self.update_cellset(cellset_id=cellset_id, values=values, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
+
+        if increment:
+            current_values = self.extract_cellset_values(cellset_id, delete_cellset=False, **kwargs)
+            values = (x + (y or None) for x, y in zip(values, current_values))
+
+        return self.update_cellset(cellset_id=cellset_id, values=values, sandbox_name=sandbox_name, **kwargs)
 
     @tidy_cellset
-    def update_cellset(self, cellset_id: str, values: List, **kwargs) -> Response:
+    def update_cellset(self, cellset_id: str, values: Iterable, sandbox_name: str = None, **kwargs) -> Response:
         """ Write values into cellset
 
         Number of values must match the number of cells in the cellset
 
-        :param cellset_id: 
+        :param cellset_id:
         :param values: iterable with Numeric and String values
-        :return: 
+        :param sandbox_name: str
+        :return:
         """
-        request = format_url("/api/v1/Cellsets('{}')/Cells", cellset_id)
+        url = format_url("/api/v1/Cellsets('{}')/Cells", cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         data = []
         for o, value in enumerate(values):
             data.append({
                 "Ordinal": o,
                 "Value": value
             })
-        return self._rest.PATCH(request, json.dumps(data, ensure_ascii=False), **kwargs)
+        return self._rest.PATCH(url, json.dumps(data, ensure_ascii=False), **kwargs)
 
-    def execute_mdx(self, mdx: str, cell_properties: List[str] = None, top: int = None,
-                    skip_contexts: bool = False, **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
+    def execute_mdx(self, mdx: str, cell_properties: List[str] = None, top: int = None, skip_contexts: bool = False,
+                    skip: int = None, skip_zeros: bool = False, skip_consolidated_cells: bool = False,
+                    skip_rule_derived_cells: bool = False, sandbox_name: str = None,
+                    **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ Execute MDX and return the cells with their properties
 
         :param mdx: MDX Query, as string
-        :param cell_properties: properties to be queried from the cell. E.g. Value, Ordinal, RuleDerived, ... 
-        :param top: integer
+        :param cell_properties: properties to be queried from the cell. E.g. Value, Ordinal, RuleDerived, ...
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
         :param skip_contexts: skip elements from titles / contexts in response
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: content in sweet concise structure.
         """
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
         return self.extract_cellset(
             cellset_id=cellset_id,
             cell_properties=cell_properties,
             top=top,
+            skip=skip,
             skip_contexts=skip_contexts,
+            skip_zeros=skip_zeros,
+            skip_consolidated_cells=skip_consolidated_cells,
+            skip_rule_derived_cells=skip_rule_derived_cells,
             delete_cellset=True,
+            sandbox_name=sandbox_name,
             **kwargs)
 
-    def execute_view(self, cube_name: str, view_name: str, cell_properties: Iterable[str] = None, private: bool = False,
-                     top: int = None, skip_contexts: bool = False,
+    def execute_view(self, cube_name: str, view_name: str, private: bool = False, cell_properties: Iterable[str] = None,
+                     top: int = None, skip_contexts: bool = False, skip: int = None, skip_zeros: bool = False,
+                     skip_consolidated_cells: bool = False, skip_rule_derived_cells: bool = False,
+                     sandbox_name: str = None,
                      **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ get view content as dictionary with sweet and concise structure.
             Works on NativeView and MDXView !
 
-        :param cube_name: String
-        :param view_name: String
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
         :param cell_properties: List, cell properties: [Values, Status, HasPicklist, etc.]
         :param private: Boolean
         :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
         :param skip_contexts: skip elements from titles / contexts in response
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
 
         :return: Dictionary : {([dim1].[elem1], [dim2][elem6]): {'Value':3127.312, 'Ordinal':12}   ....  }
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
         return self.extract_cellset(
             cellset_id=cellset_id,
             cell_properties=cell_properties,
             top=top,
+            skip=skip,
             skip_contexts=skip_contexts,
+            skip_zeros=skip_zeros,
+            skip_consolidated_cells=skip_consolidated_cells,
+            skip_rule_derived_cells=skip_rule_derived_cells,
             delete_cellset=True,
+            sandbox_name=sandbox_name,
             **kwargs)
 
     def execute_mdx_raw(
@@ -330,6 +655,11 @@ class CellService(ObjectService):
             member_properties: Iterable[str] = None,
             top: int = None,
             skip_contexts: bool = False,
+            skip: int = None,
+            skip_zeros: bool = False,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            sandbox_name: str = None,
             **kwargs) -> Dict:
         """ Execute MDX and return the raw data from TM1
 
@@ -338,18 +668,28 @@ class CellService(ObjectService):
         :param elem_properties: List of properties to be queried from the elements. E.g. ['Name','Attributes', ...]
         :param member_properties: List of properties to be queried from the members. E.g. ['Name','Attributes', ...]
         :param top: Integer limiting the number of cells and the number or rows returned
+        :param skip: Integer limiting the number of cells and the number or rows returned
         :param skip_contexts: skip elements from titles / contexts in response
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Raw format from TM1.
         """
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
         return self.extract_cellset_raw(
             cellset_id=cellset_id,
             cell_properties=cell_properties,
             elem_properties=elem_properties,
             member_properties=member_properties,
             top=top,
+            skip=skip,
             delete_cellset=True,
             skip_contexts=skip_contexts,
+            skip_zeros=skip_zeros,
+            skip_consolidated_cells=skip_consolidated_cells,
+            skip_rule_derived_cells=skip_rule_derived_cells,
+            sandbox_name=sandbox_name,
             **kwargs)
 
     def execute_view_raw(
@@ -362,8 +702,14 @@ class CellService(ObjectService):
             member_properties: Iterable[str] = None,
             top: int = None,
             skip_contexts: bool = False,
+            skip: int = None,
+            skip_zeros: bool = False,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            sandbox_name: str = None,
             **kwargs) -> Dict:
         """ Execute a cube view and return the raw data from TM1
+
 
         :param cube_name: String, name of the cube
         :param view_name: String, name of the view
@@ -373,193 +719,379 @@ class CellService(ObjectService):
         :param member_properties: List of properties to be queried from the members. E.g. ['Name','Attributes', ...]
         :param top: Integer limiting the number of cells and the number or rows returned
         :param skip_contexts: skip elements from titles / contexts in response
+        :param skip: Integer limiting the number of cells and the number or rows returned
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Raw format from TM1.
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
         return self.extract_cellset_raw(
             cellset_id=cellset_id,
             cell_properties=cell_properties,
             elem_properties=elem_properties,
             member_properties=member_properties,
             top=top,
+            skip=skip,
             skip_contexts=skip_contexts,
+            skip_zeros=skip_zeros,
+            skip_rule_derived_cells=skip_rule_derived_cells,
+            skip_consolidated_cells=skip_consolidated_cells,
             delete_cellset=True,
+            sandbox_name=sandbox_name,
             **kwargs)
 
-    def execute_mdx_values(self, mdx: str, **kwargs):
-        """ Optimized for performance. Query only raw cell values. 
+    def execute_mdx_values(self, mdx: str, sandbox_name: str = None, **kwargs) -> List[Union[str, float]]:
+        """ Optimized for performance. Query only raw cell values.
         Coordinates are omitted !
 
         :param mdx: a valid MDX Query
-        :return: Generator of cell values
+        :param sandbox_name: str
+        :return: List of cell values
         """
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
-        return self.extract_cellset_values(cellset_id, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_values(cellset_id, delete_cellset=True, sandbox_name=sandbox_name, **kwargs)
 
-    def execute_view_values(self, cube_name: str, view_name: str, private: bool = False, **kwargs):
+    def execute_view_values(self, cube_name: str, view_name: str, private: bool = False, sandbox_name: str = None,
+                            **kwargs) -> List[Union[str, float]]:
         """ Execute view and retrieve only the cell values
 
-        :param cube_name:
-        :param view_name:
-        :param private:
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
-        return self.extract_cellset_values(cellset_id, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_values(cellset_id, delete_cellset=True, sandbox_name=sandbox_name, **kwargs)
 
-    def execute_mdx_rows_and_values(self, mdx: str, element_unique_names: bool = True,
+    def execute_mdx_rows_and_values(self, mdx: str, element_unique_names: bool = True, sandbox_name: str = None,
                                     **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ Execute MDX and retrieve row element names and values in a case and space insensitive dictionary
 
         :param mdx:
         :param element_unique_names:
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
-        cellset_id = self.create_cellset(mdx=mdx, **kwargs)
-        return self.extract_cellset_rows_and_values(cellset_id, element_unique_names, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_rows_and_values(cellset_id, element_unique_names, delete_cellset=True,
+                                                    sandbox_name=sandbox_name, **kwargs)
 
     def execute_view_rows_and_values(self, cube_name: str, view_name: str, private: bool = False,
-                                     element_unique_names: bool = True, **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
+                                     element_unique_names: bool = True, sandbox_name: str = None,
+                                     **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ Execute cube view and retrieve row element names and values in a case and space insensitive dictionary
 
-        :param cube_name:
-        :param view_name:
-        :param private:
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
         :param element_unique_names:
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
-        return self.extract_cellset_rows_and_values(cellset_id, element_unique_names, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_rows_and_values(cellset_id, element_unique_names, delete_cellset=True,
+                                                    sandbox_name=sandbox_name, **kwargs)
 
-    def execute_mdx_csv(self, mdx: str, **kwargs) -> str:
-        """ Optimized for performance. Get csv string of coordinates and values. 
-        Context dimensions are omitted !
-        Cells with Zero/null are omitted !
+    def execute_mdx_csv(self, mdx: str, top: int = None, skip: int = None, skip_zeros: bool = True,
+                        skip_consolidated_cells: bool = False, skip_rule_derived_cells: bool = False,
+                        line_separator: str = "\r\n", value_separator: str = ",", sandbox_name: str = None,
+                        **kwargs) -> str:
+        """ Optimized for performance. Get csv string of coordinates and values.
 
-        :param mdx: Valid MDX Query 
+        :param mdx: Valid MDX Query
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param line_separator:
+        :param value_separator:
+        :param sandbox_name: str
         :return: String
         """
-        cellset_id = self.create_cellset(mdx, **kwargs)
-        return self.extract_cellset_csv(cellset_id=cellset_id, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset(mdx, sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_csv(cellset_id=cellset_id, top=top, skip=skip, skip_zeros=skip_zeros,
+                                        skip_consolidated_cells=skip_consolidated_cells,
+                                        skip_rule_derived_cells=skip_rule_derived_cells, line_separator=line_separator,
+                                        value_separator=value_separator, sandbox_name=sandbox_name, **kwargs)
 
-    def execute_view_csv(self, cube_name: str, view_name: str, private: bool = False, **kwargs) -> str:
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private)
-        return self.extract_cellset_csv(cellset_id=cellset_id, delete_cellset=True, **kwargs)
+    def execute_view_csv(self, cube_name: str, view_name: str, private: bool = False, top: int = None, skip: int = None,
+                         skip_zeros: bool = True, skip_consolidated_cells: bool = False,
+                         skip_rule_derived_cells: bool = False,
+                         line_separator: str = "\r\n", value_separator: str = ",", sandbox_name: str = None,
+                         **kwargs) -> str:
+        """ Optimized for performance. Get csv string of coordinates and values.
 
-    def execute_mdx_dataframe(self, mdx: str, **kwargs) -> pd.DataFrame:
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param line_separator:
+        :param value_separator:
+        :param sandbox_name: str
+        :return: String
+        """
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name)
+        return self.extract_cellset_csv(cellset_id=cellset_id, skip_zeros=skip_zeros, top=top, skip=skip,
+                                        skip_consolidated_cells=skip_consolidated_cells,
+                                        skip_rule_derived_cells=skip_rule_derived_cells, line_separator=line_separator,
+                                        value_separator=value_separator, sandbox_name=sandbox_name, **kwargs)
+
+    def execute_mdx_elements_value_dict(self, mdx: str, top: int = None, skip: int = None, skip_zeros: bool = True,
+                                        skip_consolidated_cells: bool = False, skip_rule_derived_cells: bool = False,
+                                        element_separator: str = "|",
+                                        sandbox_name: str = None, **kwargs) -> CaseAndSpaceInsensitiveDict:
+        """ Optimized for performance. Get Dict from MDX Query.
+        :param mdx: Valid MDX Query
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param element_separator: separator for the dimension element combination
+        :param sandbox_name: str
+        :return: CaseAndSpaceInsensitiveDict {'2020|Jan|Sales': 2000, '2020|Feb|Sales': 3000}
+        """
+        lines = self.execute_mdx_csv(mdx=mdx, top=top, skip=skip, skip_zeros=skip_zeros,
+                                     skip_consolidated_cells=skip_consolidated_cells,
+                                     skip_rule_derived_cells=skip_rule_derived_cells,
+                                     value_separator=element_separator,
+                                     sandbox_name=sandbox_name, **kwargs)
+        elements_value_dict = CaseAndSpaceInsensitiveDict()
+        for entries in lines.split("\r\n")[1:]:
+            elements_value_dict[
+                element_separator.join(entries.split(element_separator)[:-1])] = entries.split(element_separator)[-1]
+        return elements_value_dict
+
+    @require_pandas
+    def execute_mdx_dataframe(self, mdx: str, top: int = None, skip: int = None, skip_zeros: bool = True,
+                              skip_consolidated_cells: bool = False, skip_rule_derived_cells: bool = False,
+                              sandbox_name: str = None,
+                              **kwargs) -> 'pd.DataFrame':
         """ Optimized for performance. Get Pandas DataFrame from MDX Query.
-
-        Context dimensions are omitted in the resulting Dataframe !
-        Cells with Zero/null are omitted !
 
         Takes all arguments from the pandas.read_csv method:
         https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.read_csv.html
 
         :param mdx: Valid MDX Query
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Pandas Dataframe
         """
-        cellset_id = self.create_cellset(mdx, **kwargs)
-        return self.extract_cellset_dataframe(cellset_id, **kwargs)
+        cellset_id = self.create_cellset(mdx, sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_dataframe(cellset_id, top=top, skip=skip, skip_zeros=skip_zeros,
+                                              skip_consolidated_cells=skip_consolidated_cells,
+                                              skip_rule_derived_cells=skip_rule_derived_cells,
+                                              sandbox_name=sandbox_name, **kwargs)
 
-    def execute_view_dataframe_pivot(self, cube_name: str, view_name: str, private: bool = False, dropna: bool = False,
-                                     fill_value: bool = None, **kwargs) -> pd.DataFrame:
-        """ Execute a cube view to get a pandas pivot dataframe, in the shape of the cube view
+    @require_pandas
+    def execute_mdx_dataframe_shaped(self, mdx: str, sandbox_name: str = None, **kwargs) -> 'pd.DataFrame':
+        """ Retrieves data from cube in the shape of the query.
+        Dimensions on rows can be stacked. One dimension must be placed on columns. Title selections are ignored.
+
+        :param mdx:
+        :param sandbox_name: str
+        :param kwargs:
+        :return:
+        """
+        cellset_id = self.create_cellset(mdx, sandbox_name=sandbox_name)
+        return self.extract_cellset_dataframe_shaped(cellset_id, delete_cellset=True, sandbox_name=sandbox_name,
+                                                     **kwargs)
+
+    @require_pandas
+    def execute_view_dataframe_shaped(self, cube_name: str, view_name: str, private: bool = False,
+                                      sandbox_name: str = None,
+                                      **kwargs) -> 'pd.DataFrame':
+        """ Retrieves data from cube in the shape of the query.
+        Dimensions on rows can be stacked. One dimension must be placed on columns. Title selections are ignored.
 
         :param cube_name:
         :param view_name:
         :param private:
-        :param dropna:
-        :param fill_value:
+        :param sandbox_name: str
+        :param kwargs:
         :return:
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name, view_name, private, sandbox_name=sandbox_name)
+        return self.extract_cellset_dataframe_shaped(cellset_id, delete_cellset=True, sandbox_name=sandbox_name,
+                                                     **kwargs)
+
+    @require_pandas
+    def execute_view_dataframe_pivot(self, cube_name: str, view_name: str, private: bool = False, dropna: bool = False,
+                                     fill_value: bool = None, sandbox_name: str = None, **kwargs) -> 'pd.DataFrame':
+        """ Execute a cube view to get a pandas pivot dataframe, in the shape of the cube view
+
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
+        :param dropna:
+        :param fill_value:
+        :param sandbox_name: str
+        :return:
+        """
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
         return self.extract_cellset_dataframe_pivot(
             cellset_id=cellset_id,
             dropna=dropna,
             fill_value=fill_value,
+            sandbox_name=sandbox_name,
             **kwargs)
 
-    def execute_mdx_dataframe_pivot(self, mdx: str, dropna: bool = False, fill_value: bool = None) -> pd.DataFrame:
+    @require_pandas
+    def execute_mdx_dataframe_pivot(self, mdx: str, dropna: bool = False, fill_value: bool = None,
+                                    sandbox_name: str = None) -> 'pd.DataFrame':
         """ Execute MDX Query to get a pandas pivot data frame in the shape as specified in the Query
 
         :param mdx:
         :param dropna:
         :param fill_value:
+        :param sandbox_name: str
         :return:
         """
-        cellset_id = self.create_cellset(mdx=mdx)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name)
         return self.extract_cellset_dataframe_pivot(
             cellset_id=cellset_id,
             dropna=dropna,
-            fill_value=fill_value)
+            fill_value=fill_value,
+            sandbox_name=sandbox_name)
 
-    def execute_view_dataframe(self, cube_name: str, view_name: str, private: bool = False, **kwargs) -> pd.DataFrame:
-        """ Optimized for performance. Get Pandas DataFrame from an existing Cube View 
+    def execute_mdx_cellcount(self, mdx: str, sandbox_name: str = None, **kwargs) -> int:
+        """ Execute MDX in order to understand how many cells are in a cellset.
+        Only return number of cells in the cellset. FAST!
+
+        :param mdx: MDX Query, as string
+        :param sandbox_name: str
+        :return: Number of Cells in the CellSet
+        """
+        cellset_id = self.create_cellset(mdx, sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_cellcount(cellset_id, delete_cellset=True, sandbox_name=sandbox_name, **kwargs)
+
+    def execute_view_elements_value_dict(self, cube_name: str, view_name: str, private: bool = False,
+                                         top: int = None, skip: int = None, skip_zeros: bool = True,
+                                         skip_consolidated_cells: bool = False, skip_rule_derived_cells: bool = False,
+                                         element_separator: str = "|", sandbox_name: str = None,
+                                         **kwargs) -> CaseAndSpaceInsensitiveDict:
+        """ Optimized for performance. Get a Dict(tuple, value) from an existing Cube View
+        Context dimensions are omitted in the resulting Dataframe !
+        Cells with Zero/null are omitted by default, but still configurable!
+
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param element_separator: separator for the dimension element combination
+        :param sandbox_name: str
+        :return: CaseAndSpaceInsensitiveDict {'2020|Jan|Sales': 2000, '2020|Feb|Sales': 3000}
+        """
+        lines = self.execute_view_csv(cube_name=cube_name, view_name=view_name, private=private, top=top, skip=skip,
+                                      skip_zeros=skip_zeros, skip_consolidated_cells=skip_consolidated_cells,
+                                      skip_rule_derived_cells=skip_rule_derived_cells,
+                                      value_separator=element_separator, sandbox_name=sandbox_name, **kwargs)
+        elements_value_dict = CaseAndSpaceInsensitiveDict()
+        for entries in lines.split("\r\n")[1:]:
+            elements_value_dict[
+                element_separator.join(entries.split(element_separator)[:-1])] = entries.split(element_separator)[-1]
+        return elements_value_dict
+
+    @require_pandas
+    def execute_view_dataframe(self, cube_name: str, view_name: str, private: bool = False, top: int = None,
+                               skip: int = None, skip_zeros: bool = True, skip_consolidated_cells: bool = False,
+                               skip_rule_derived_cells: bool = False, sandbox_name: str = None,
+                               **kwargs) -> 'pd.DataFrame':
+        """ Optimized for performance. Get Pandas DataFrame from an existing Cube View
         Context dimensions are omitted in the resulting Dataframe !
         Cells with Zero/null are omitted !
 
         Takes all arguments from the pandas.read_csv method:
         https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.read_csv.html
 
-        :param cube_name: Name of the 
-        :param view_name: 
-        :param private: 
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Pandas Dataframe
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
-        return self.extract_cellset_dataframe(cellset_id, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_dataframe(cellset_id, top=top, skip=skip, skip_zeros=skip_zeros,
+                                              skip_consolidated_cells=skip_consolidated_cells,
+                                              skip_rule_derived_cells=skip_rule_derived_cells,
+                                              sandbox_name=sandbox_name, **kwargs)
 
-    def execute_mdx_cellcount(self, mdx: str, **kwargs) -> int:
-        """ Execute MDX in order to understand how many cells are in a cellset.
-        Only return number of cells in the cellset. FAST!
-
-        :param mdx: MDX Query, as string
-        :return: Number of Cells in the CellSet
-        """
-        cellset_id = self.create_cellset(mdx, **kwargs)
-        return self.extract_cellset_cellcount(cellset_id, delete_cellset=True, **kwargs)
-
-    def execute_view_cellcount(self, cube_name: str, view_name: str, private: bool = False, **kwargs) -> int:
+    def execute_view_cellcount(self, cube_name: str, view_name: str, private: bool = False, sandbox_name: str = None,
+                               **kwargs) -> int:
         """ Execute cube view in order to understand how many cells are in a cellset.
         Only return number of cells in the cellset. FAST!
-        
-        :param cube_name: cube name
-        :param view_name: view name
+
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
         :param private: True (private) or False (public)
-        :return: 
+        :param sandbox_name: str
+        :return:
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
-        return self.extract_cellset_cellcount(cellset_id, delete_cellset=True, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
+        return self.extract_cellset_cellcount(cellset_id, delete_cellset=True, sandbox_name=sandbox_name, **kwargs)
 
     def execute_mdx_rows_and_values_string_set(
             self,
             mdx: str,
             exclude_empty_cells: bool = True,
+            sandbox_name: str = None,
             **kwargs) -> CaseAndSpaceInsensitiveSet:
         """ Retrieve row element names and **string** cell values in a case and space insensitive set
 
         :param exclude_empty_cells:
         :param mdx:
+        :param sandbox_name: str
         :return:
         """
-        rows_and_values = self.execute_mdx_rows_and_values(mdx, element_unique_names=False, **kwargs)
+        rows_and_values = self.execute_mdx_rows_and_values(mdx, element_unique_names=False, sandbox_name=sandbox_name,
+                                                           **kwargs)
         return self._extract_string_set_from_rows_and_values(rows_and_values, exclude_empty_cells)
 
     def execute_view_rows_and_values_string_set(self, cube_name: str, view_name: str, private: bool = False,
-                                                exclude_empty_cells: bool = True,
+                                                exclude_empty_cells: bool = True, sandbox_name: str = None,
                                                 **kwargs) -> CaseAndSpaceInsensitiveSet:
         """ Retrieve row element names and **string** cell values in a case and space insensitive set
 
-        :param cube_name:
-        :param view_name:
-        :param private:
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
         :param exclude_empty_cells:
+        :param sandbox_name: str
         :return:
         """
-        rows_and_values = self.execute_view_rows_and_values(cube_name, view_name, private, False, **kwargs)
+        rows_and_values = self.execute_view_rows_and_values(cube_name, view_name, private, False,
+                                                            sandbox_name=sandbox_name, **kwargs)
         return self._extract_string_set_from_rows_and_values(rows_and_values, exclude_empty_cells)
 
     def execute_mdx_ui_dygraph(
@@ -567,8 +1099,10 @@ class CellService(ObjectService):
             mdx: str,
             elem_properties: Iterable[str] = None,
             member_properties: Iterable[str] = None,
-            value_precision: Iterable[str] = 2,
+            value_precision: int = 2,
             top: int = None,
+            skip: int = None,
+            sandbox_name: str = None,
             **kwargs) -> Dict:
         """ Execute MDX get dygraph dictionary
         Useful for grids or charting libraries that want an array of cell values per column
@@ -586,20 +1120,24 @@ class CellService(ObjectService):
                     ['Q3-2004', 14502421.63, 10466934.096533755],
                     ['Q4-2004', 14321501.940000001, 10333095.839474997]]
             },
-        :param top:
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
         :param mdx: String, valid MDX Query
         :param elem_properties: List of properties to be queried from the elements. E.g. ['UniqueName','Attributes']
         :param member_properties: List of properties to be queried from the members. E.g. ['UniqueName','Attributes']
         :param value_precision: Integer (optional) specifying number of decimal places to return
+        :param sandbox_name: str
         :return: dict: { titles: [], headers: [axis][], cells: { Page0: [ [column name, column values], [], ... ], ...}}
         """
-        cellset_id = self.create_cellset(mdx)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name)
         data = self.extract_cellset_raw(cellset_id=cellset_id,
                                         cell_properties=["Value"],
                                         elem_properties=elem_properties,
                                         member_properties=list(set(member_properties or []) | {"Name"}),
                                         top=top,
+                                        skip=skip,
                                         delete_cellset=True,
+                                        sandbox_name=sandbox_name,
                                         **kwargs)
         return Utils.build_ui_dygraph_arrays_from_cellset(raw_cellset_as_dict=data, value_precision=value_precision)
 
@@ -612,6 +1150,8 @@ class CellService(ObjectService):
             member_properties: Iterable[str] = None,
             value_precision: int = 2,
             top: int = None,
+            skip: int = None,
+            sandbox_name: str = None,
             **kwargs):
         """
         Useful for grids or charting libraries that want an array of cell values per row.
@@ -639,22 +1179,27 @@ class CellService(ObjectService):
                                  14321501.940000001]}
             },
 
-        :param top:
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
         :param cube_name: cube name
         :param view_name: view name
         :param private: True (private) or False (public)
         :param elem_properties: List of properties to be queried from the elements. E.g. ['UniqueName','Attributes']
         :param member_properties: List of properties to be queried from the members. E.g. ['UniqueName','Attributes']
         :param value_precision: Integer (optional) specifying number of decimal places to return
+        :param sandbox_name: str
         :return:
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
         data = self.extract_cellset_raw(cellset_id=cellset_id,
                                         cell_properties=["Value"],
                                         elem_properties=elem_properties,
                                         member_properties=list(set(member_properties or []) | {"Name"}),
                                         top=top,
+                                        skip=skip,
                                         delete_cellset=True,
+                                        sandbox_name=sandbox_name,
                                         **kwargs)
         return Utils.build_ui_dygraph_arrays_from_cellset(raw_cellset_as_dict=data, value_precision=value_precision)
 
@@ -665,6 +1210,8 @@ class CellService(ObjectService):
             member_properties: Iterable[str] = None,
             value_precision: int = 2,
             top: int = None,
+            skip: int = None,
+            sandbox_name: str = None,
             **kwargs):
         """
         Useful for grids or charting libraries that want an array of cell values per row.
@@ -692,20 +1239,24 @@ class CellService(ObjectService):
                                  14321501.940000001]}
             },
 
-        :param top:
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
         :param mdx: a valid MDX Query
         :param elem_properties: List of properties to be queried from the elements. E.g. ['UniqueName','Attributes']
         :param member_properties: List of properties to be queried from the members. E.g. ['UniqueName','Attributes']
         :param value_precision: Integer (optional) specifying number of decimal places to return
+        :param sandbox_name: str
         :return: dict :{ titles: [], headers: [axis][], cells:{ Page0:{ Row0:{ [row values], Row1: [], ...}, ...}, ...}}
         """
-        cellset_id = self.create_cellset(mdx, **kwargs)
+        cellset_id = self.create_cellset(mdx=mdx, sandbox_name=sandbox_name, **kwargs)
         data = self.extract_cellset_raw(cellset_id=cellset_id,
                                         cell_properties=["Value"],
                                         elem_properties=elem_properties,
                                         member_properties=list(set(member_properties or []) | {"Name"}),
                                         top=top,
+                                        skip=skip,
                                         delete_cellset=True,
+                                        sandbox_name=sandbox_name,
                                         **kwargs)
         return Utils.build_ui_arrays_from_cellset(raw_cellset_as_dict=data, value_precision=value_precision)
 
@@ -718,6 +1269,8 @@ class CellService(ObjectService):
             member_properties: Iterable[str] = None,
             value_precision: int = 2,
             top: int = None,
+            skip: int = None,
+            sandbox_name: str = None,
             **kwargs):
         """
         Useful for grids or charting libraries that want an array of cell values per row.
@@ -745,22 +1298,27 @@ class CellService(ObjectService):
                                  14321501.940000001]}
             },
 
-        :param top:
-        :param cube_name: cube name
-        :param view_name: view name
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
         :param private: True (private) or False (public)
         :param elem_properties: List of properties to be queried from the elements. E.g. ['UniqueName','Attributes']
         :param member_properties: List properties to be queried from the member. E.g. ['Name', 'UniqueName']
         :param value_precision: Integer (optional) specifying number of decimal places to return
+        :param sandbox_name: str
         :return: dict :{ titles: [], headers: [axis][], cells:{ Page0:{ Row0: {[row values], Row1: [], ...}, ...}, ...}}
         """
-        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private, **kwargs)
+        cellset_id = self.create_cellset_from_view(cube_name=cube_name, view_name=view_name, private=private,
+                                                   sandbox_name=sandbox_name, **kwargs)
         data = self.extract_cellset_raw(cellset_id=cellset_id,
                                         cell_properties=["Value"],
                                         elem_properties=elem_properties,
                                         member_properties=list(set(member_properties or []) | {"Name"}),
                                         top=top,
+                                        skip=skip,
                                         delete_cellset=True,
+                                        sandbox_name=sandbox_name,
                                         **kwargs)
         return Utils.build_ui_arrays_from_cellset(raw_cellset_as_dict=data, value_precision=value_precision)
 
@@ -772,7 +1330,12 @@ class CellService(ObjectService):
             elem_properties: Iterable[str] = None,
             member_properties: Iterable[str] = None,
             top: int = None,
+            skip: int = None,
             skip_contexts: bool = False,
+            skip_zeros: bool = False,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            sandbox_name: str = None,
             **kwargs) -> Dict:
         """ Extract full cellset data and return the raw data from TM1
 
@@ -781,11 +1344,28 @@ class CellService(ObjectService):
         :param elem_properties: List of properties to be queried from elements. E.g. ['UniqueName','Attributes', ...]
         :param member_properties: List properties to be queried from the member. E.g. ['Name', 'UniqueName']
         :param top: Integer limiting the number of cells and the number or rows returned
+        :param skip: Integer limiting the number of cells and the number or rows returned
         :param skip_contexts:
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Raw format from TM1.
         """
         if not cell_properties:
             cell_properties = ['Value']
+
+        if skip_rule_derived_cells:
+            cell_properties.append("RuleDerived")
+            # necessary due to bug in TM1 11.8: If only RuleDerived is retrieved it occasionally produces wrong results
+            cell_properties.append("Updateable")
+
+        if skip_consolidated_cells:
+            cell_properties.append("Consolidated")
+
+        if skip or skip_zeros or skip_rule_derived_cells or skip_consolidated_cells:
+            if 'Ordinal' not in cell_properties:
+                cell_properties.append('Ordinal')
 
         # select Name property if member_properties is None or empty.
         # Necessary, as tm1 default behaviour is to return all properties if no $select is specified in the request.
@@ -800,46 +1380,66 @@ class CellService(ObjectService):
 
         filter_axis = "$filter=Ordinal ne 2;" if skip_contexts else ""
 
+        filter_cells = ""
+        if skip_zeros or skip_consolidated_cells or skip_rule_derived_cells:
+            filters = []
+            if skip_zeros:
+                filters.append("Value ne 0 and Value ne null")
+            if skip_consolidated_cells:
+                filters.append("Consolidated eq false")
+            if skip_rule_derived_cells:
+                filters.append("RuleDerived eq false")
+
+            filter_cells = " and ".join(filters)
+
         url = "/api/v1/Cellsets('{cellset_id}')?$expand=" \
               "Cube($select=Name;$expand=Dimensions($select=Name))," \
               "Axes({filter_axis}$expand=Tuples($expand=Members({select_member_properties}" \
-              "{expand_elem_properties}){top_rows}))," \
-              "Cells($select={cell_properties}{top_cells})" \
+              "{expand_elem_properties}{top_rows})))," \
+              "Cells($select={cell_properties}{top_cells}{skip_cells}{filter_cells})" \
             .format(cellset_id=cellset_id,
-                    top_rows=";$top={}".format(top) if top else "",
+                    top_rows=f";$top={top}" if top and not skip else "",
                     cell_properties=",".join(cell_properties),
                     filter_axis=filter_axis,
                     select_member_properties=select_member_properties,
                     expand_elem_properties=expand_elem_properties,
-                    top_cells=";$top={}".format(top) if top else "")
+                    top_cells=f";$top={top}" if top else "",
+                    skip_cells=f";$skip={skip}" if skip else "",
+                    filter_cells=f";$filter={filter_cells}" if filter_cells else "")
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url=url, **kwargs)
         return response.json()
 
     @tidy_cellset
-    def extract_cellset_values(self, cellset_id: str, **kwargs):
+    def extract_cellset_values(self, cellset_id: str, sandbox_name: str = None, **kwargs) -> List[Union[str, float]]:
         """ Extract cellset data and return only the cells and values
 
         :param cellset_id: String; ID of existing cellset
+        :param sandbox_name: str
         :return: Raw format from TM1.
         """
         url = format_url("/api/v1/Cellsets('{}')?$expand=Cells($select=Value)", cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url=url, **kwargs)
-        return (cell["Value"] for cell in response.json()["Cells"])
+        return [cell["Value"] for cell in response.json()["Cells"]]
 
     @tidy_cellset
     def extract_cellset_rows_and_values(self, cellset_id: str, element_unique_names: bool = True,
+                                        sandbox_name: str = None,
                                         **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ Retrieve row element names and values in a case and space insensitive dictionary
 
         :param cellset_id:
         :param element_unique_names:
         :param kwargs:
+        :param sandbox_name: str
         :return:
         """
         url = "/api/v1/Cellsets('{}')?$expand=" \
               "Axes($filter=Ordinal eq 1;$expand=Tuples(" \
               "$expand=Members($select=Element;$expand=Element($select={}))))," \
               "Cells($select=Value)".format(cellset_id, "UniqueName" if element_unique_names else "Name")
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url=url, **kwargs)
         response_json = response.json()
         rows = response_json["Axes"][0]["Tuples"]
@@ -867,72 +1467,135 @@ class CellService(ObjectService):
         return result
 
     @tidy_cellset
-    def extract_cellset_composition(self, cellset_id: str, **kwargs):
+    def extract_cellset_composition(self, cellset_id: str, sandbox_name: str = None, **kwargs):
         """ Retrieve composition of dimensions on the axes in the cellset
 
         :param cellset_id:
         :param kwargs:
+        :param sandbox_name: str
         :return:
         """
         url = "/api/v1/Cellsets('{}')?$expand=" \
               "Cube($select=Name)," \
               "Axes($expand=Hierarchies($select=UniqueName))".format(cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url=url, **kwargs)
         response_json = response.json()
         cube = response_json["Cube"]["Name"]
 
         rows, titles, columns = [], [], []
-        if response_json["Axes"][0]["Hierarchies"]:
-            columns = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][0]["Hierarchies"]]
-        if response_json["Axes"][1]["Hierarchies"]:
-            rows = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][1]["Hierarchies"]]
+        if len(response_json["Axes"]) == 1:
+            if response_json["Axes"][0]["Hierarchies"]:
+                columns = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][0]["Hierarchies"]]
+        else:
+            if response_json["Axes"][0]["Hierarchies"]:
+                columns = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][0]["Hierarchies"]]
+            if response_json["Axes"][1]["Hierarchies"]:
+                rows = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][1]["Hierarchies"]]
         if len(response_json["Axes"]) > 2:
             titles = [hierarchy["UniqueName"] for hierarchy in response_json["Axes"][2]["Hierarchies"]]
         return cube, titles, rows, columns
 
     @tidy_cellset
-    def extract_cellset_cellcount(self, cellset_id: str, **kwargs) -> int:
+    def extract_cellset_cellcount(self, cellset_id: str, sandbox_name: str = None, **kwargs) -> int:
         """ Retrieve number of cells in the cellset
 
         :param cellset_id:
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
         url = "/api/v1/Cellsets('{}')/Cells/$count".format(cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url, **kwargs)
         return int(response.content)
 
-    @tidy_cellset
-    def extract_cellset_csv(self, cellset_id: str, **kwargs) -> str:
+    def extract_cellset_csv(
+            self,
+            cellset_id: str,
+            top: int = None,
+            skip: int = None,
+            skip_zeros: bool = True,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            line_separator: str = "\r\n",
+            value_separator: str = ",",
+            sandbox_name: str = None,
+            **kwargs) -> str:
         """ Execute cellset and return only the 'Content', in csv format
 
         :param cellset_id: String; ID of existing cellset
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param line_separator:
+        :param value_separator
+        :param sandbox_name: str
         :return: Raw format from TM1.
         """
-        url = "/api/v1/Cellsets('{}')/Content".format(cellset_id)
-        data = self._rest.GET(url, **kwargs)
-        return data.text
+        _, _, rows, columns = self.extract_cellset_composition(cellset_id, delete_cellset=False,
+                                                               sandbox_name=sandbox_name, **kwargs)
+        cellset_dict = self.extract_cellset_raw(cellset_id, cell_properties=["Value"], top=top, skip=skip,
+                                                skip_contexts=True, skip_zeros=skip_zeros,
+                                                skip_consolidated_cells=skip_consolidated_cells,
+                                                skip_rule_derived_cells=skip_rule_derived_cells,
+                                                delete_cellset=True, sandbox_name=sandbox_name, **kwargs)
+        return build_csv_from_cellset_dict(rows, columns, cellset_dict, line_separator=line_separator,
+                                           value_separator=value_separator, top=top)
 
-    def extract_cellset_dataframe(self, cellset_id: str, **kwargs) -> pd.DataFrame:
+    @require_pandas
+    def extract_cellset_dataframe(
+            self,
+            cellset_id: str,
+            top: int = None,
+            skip: int = None,
+            skip_zeros: bool = True,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            sandbox_name: str = None,
+            **kwargs) -> 'pd.DataFrame':
         """ Build pandas data frame from cellset_id
 
         :param cellset_id:
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :param kwargs:
         :return:
         """
-        raw_csv = self.extract_cellset_csv(cellset_id=cellset_id, delete_cellset=True, **kwargs)
+        raw_csv = self.extract_cellset_csv(cellset_id=cellset_id, top=top, skip=skip, skip_zeros=skip_zeros,
+                                           skip_rule_derived_cells=skip_rule_derived_cells,
+                                           skip_consolidated_cells=skip_consolidated_cells, value_separator='~',
+                                           sandbox_name=sandbox_name,
+                                           **kwargs)
+        if not raw_csv:
+            return pd.DataFrame()
+
         memory_file = StringIO(raw_csv)
         # make sure all element names are strings and values column is derived from data
         if 'dtype' not in kwargs:
             kwargs['dtype'] = {'Value': None, **{col: str for col in range(999)}}
-        return pd.read_csv(memory_file, sep=',', **kwargs)
+        return pd.read_csv(memory_file, sep='~', **kwargs)
 
     @tidy_cellset
-    def extract_cellset_power_bi(self, cellset_id: str, **kwargs) -> pd.DataFrame:
+    @require_pandas
+    def extract_cellset_dataframe_shaped(self, cellset_id: str, sandbox_name: str = None, **kwargs) -> 'pd.DataFrame':
+        """ Retrieves data from cellset in the shape of the query.
+        Dimensions on rows can be stacked. One dimension must be placed on columns. Title selections are ignored.
+
+        :param cellset_id
+        :param sandbox_name: str
+        """
         url = "/api/v1/Cellsets('{}')?$expand=" \
               "Axes($filter=Ordinal eq 0 or Ordinal eq 1;$expand=Tuples(" \
               "$expand=Members($select=Name)),Hierarchies($select=Name))," \
               "Cells($select=Value)".format(cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         response = self._rest.GET(url=url, **kwargs)
         response_json = response.json()
         rows = response_json["Axes"][1]["Tuples"]
@@ -947,40 +1610,49 @@ class CellService(ObjectService):
         # avoid division by zero
         if not number_rows:
             return pd.DataFrame(body, columns=headers)
+
         number_cells = len(cell_values)
         number_columns = int(number_cells / number_rows)
+
+        element_names_by_row = [tuple(member["Name"] for member in tupl["Members"])
+                                for tupl
+                                in rows]
+
+        if not number_columns:
+            return pd.DataFrame(data=element_names_by_row, columns=headers)
 
         cell_values_by_row = [cell_values[cell_counter:cell_counter + number_columns]
                               for cell_counter
                               in range(0, number_cells, number_columns)]
-        element_names_by_row = [tuple(member["Name"]
-                                      for member
-                                      in tupl["Members"])
-                                for tupl
-                                in rows]
 
         for element_tuple, cells in zip(element_names_by_row, cell_values_by_row):
             body.append(list(element_tuple) + cells)
         return pd.DataFrame(body, columns=headers, dtype=str)
 
+    @require_pandas
     def extract_cellset_dataframe_pivot(self, cellset_id: str, dropna: bool = False, fill_value: bool = False,
-                                        **kwargs) -> pd.DataFrame:
+                                        sandbox_name: str = None,
+                                        **kwargs) -> 'pd.DataFrame':
         """ Extract a pivot table (pandas dataframe) from a cellset in TM1
 
         :param cellset_id:
         :param dropna:
         :param fill_value:
         :param kwargs:
+        :param sandbox_name: str
         :return:
         """
+
         data = self.extract_cellset(
             cellset_id=cellset_id,
             delete_cellset=False,
+            sandbox_name=sandbox_name,
             **kwargs)
 
         cube, titles, rows, columns = self.extract_cellset_composition(
             cellset_id=cellset_id,
             delete_cellset=True,
+            sandbox_name=sandbox_name,
             **kwargs)
 
         df = build_pandas_dataframe_from_cellset(data, multiindex=False)
@@ -999,8 +1671,13 @@ class CellService(ObjectService):
             cellset_id: str,
             cell_properties: Iterable[str] = None,
             top: int = None,
+            skip: int = None,
             delete_cellset: bool = True,
             skip_contexts: bool = False,
+            skip_zeros: bool = False,
+            skip_consolidated_cells: bool = False,
+            skip_rule_derived_cells: bool = False,
+            sandbox_name: str = None,
             **kwargs) -> CaseAndSpaceInsensitiveTuplesDict:
         """ Execute cellset and return the cells with their properties
 
@@ -1008,7 +1685,12 @@ class CellService(ObjectService):
         :param delete_cellset:
         :param cellset_id:
         :param cell_properties: properties to be queried from the cell. E.g. Value, Ordinal, RuleDerived, ...
-        :param top: integer
+        :param top: Int, number of cells to return (counting from top)
+        :param skip: Int, number of cells to skip (counting from top)
+        :param skip_zeros: skip zeros in cellset (irrespective of zero suppression in MDX / view)
+        :param skip_consolidated_cells: skip consolidated cells in cellset
+        :param skip_rule_derived_cells: skip rule derived cells in cellset
+        :param sandbox_name: str
         :return: Content in sweet consice strcuture.
         """
         if not cell_properties:
@@ -1020,21 +1702,28 @@ class CellService(ObjectService):
             elem_properties=['UniqueName'],
             member_properties=['UniqueName'],
             top=top,
+            skip=skip,
             skip_contexts=skip_contexts,
             delete_cellset=delete_cellset,
+            skip_zeros=skip_zeros,
+            skip_consolidated_cells=skip_consolidated_cells,
+            skip_rule_derived_cells=skip_rule_derived_cells,
+            sandbox_name=sandbox_name,
             **kwargs)
 
-        return Utils.build_content_from_cellset(
+        return Utils.build_content_from_cellset_dict(
             raw_cellset_as_dict=raw_cellset,
             top=top)
 
-    def create_cellset(self, mdx: str, **kwargs) -> str:
+    def create_cellset(self, mdx: str, sandbox_name: str = None, **kwargs) -> str:
         """ Execute MDX in order to create cellset at server. return the cellset-id
 
         :param mdx: MDX Query, as string
+        :param sandbox_name: str
         :return:
         """
         url = '/api/v1/ExecuteMDX'
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         data = {
             'MDX': mdx
         }
@@ -1042,29 +1731,41 @@ class CellService(ObjectService):
         cellset_id = response.json()['ID']
         return cellset_id
 
-    def create_cellset_from_view(self, cube_name: str, view_name: str, private: bool, **kwargs) -> str:
+    def create_cellset_from_view(self, cube_name: str, view_name: str, private: bool, sandbox_name: str = None,
+                                 **kwargs) -> str:
         """ create cellset from a cube view. return the cellset-id
 
-        :param cube_name:
-        :param view_name:
-        :param private:
+        :param cube_name: String, name of the cube
+        :param view_name: String, name of the view
+        :param private: True (private) or False (public)
         :param kwargs:
+        :param sandbox_name: str
         :return:
         """
         url = format_url("/api/v1/Cubes('{cube_name}')/{views}('{view_name}')/tm1.Execute",
                          cube_name=cube_name,
                          views='PrivateViews' if private else 'Views',
                          view_name=view_name)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         return self._rest.POST(url=url, **kwargs).json()['ID']
 
-    def delete_cellset(self, cellset_id: str, **kwargs) -> Response:
+    def delete_cellset(self, cellset_id: str, sandbox_name: str = None, **kwargs) -> Response:
         """ Delete a cellset
 
         :param cellset_id:
+        :param sandbox_name: str
         :return:
         """
         url = "/api/v1/Cellsets('{}')".format(cellset_id)
+        url = add_url_parameters(url, **{"!sandbox": sandbox_name})
         return self._rest.DELETE(url, **kwargs)
+
+    def transaction_log_is_active(self, cube_name: str) -> bool:
+        mdx = f"""
+        SELECT {{[}}Cubes].[{cube_name}]}} ON 0, {{[}}CubeProperties].[LOGGING]}} ON 1 FROM [}}CubeProperties]
+        """
+        values = self.execute_mdx_values(mdx)
+        return case_and_space_insensitive_equals(values[0], "YES")
 
     def deactivate_transactionlog(self, *args: str, **kwargs) -> Response:
         """ Deactivate Transactionlog for one or many cubes
@@ -1110,7 +1811,7 @@ class CellService(ObjectService):
             PendingDeprecationWarning
         )
         warnings.simplefilter('default', PendingDeprecationWarning)
-        return self.execute_view(cube_name, view_name, cell_properties, private, top)
+        return self.execute_view(cube_name, view_name, private, cell_properties, top)
 
     @staticmethod
     def _extract_string_set_from_rows_and_values(
