@@ -14,13 +14,42 @@ class TM1pyTimeout(Exception):
         return f"Timeout after {self.timeout} seconds for '{self.method}' request with url :'{self.url}'"
 
 
+class TM1pyVersionException(Exception):
+    def __init__(self, function: str, required_version):
+        self.function = function
+        self.required_version = required_version
+
+    def __str__(self):
+        return f"Function '{self.function}' requires TM1 server version >= '{self.required_version}'"
+
+
+class TM1pyNotAdminException(Exception):
+    def __init__(self, function: str):
+        self.function = function
+
+    def __str__(self):
+        return f"Function '{self.function}' requires admin permissions"
+
+
 class TM1pyException(Exception):
     """ The default exception for TM1py
 
     """
 
+    def __init__(self, message):
+        self.message = message
+
+    def __str__(self):
+        return self.message
+
+
+class TM1pyRestException(TM1pyException):
+    """ Exception for failing REST operations
+
+    """
+
     def __init__(self, response: str, status_code: int, reason: str, headers: Mapping):
-        self._response = response
+        super(TM1pyRestException, self).__init__(response)
         self._status_code = status_code
         self._reason = reason
         self._headers = headers
@@ -31,11 +60,11 @@ class TM1pyException(Exception):
 
     @property
     def reason(self):
-        return self.reason
+        return self._reason
 
     @property
     def response(self):
-        return self._response
+        return self.message
 
     @property
     def headers(self):
@@ -43,7 +72,7 @@ class TM1pyException(Exception):
 
     def __str__(self):
         return "Text: {} Status Code: {} Reason: {} Headers: {}".format(
-            self._response,
+            self.message,
             self._status_code,
             self._reason,
             self._headers)
