@@ -1,1 +1,1 @@
-from TM1py.Exceptions.Exceptions import TM1pyException
+from TM1py.Exceptions.Exceptions import TM1pyRestException, TM1pyException, TM1pyTimeout
