@@ -2,11 +2,12 @@
 
 import collections
 import json
-from typing import Iterable, Optional, List, Dict
 from enum import Enum
+from typing import Iterable, Optional, List, Dict, Union
 
 from TM1py.Objects.TM1Object import TM1Object
 from TM1py.Utils.Utils import CaseAndSpaceInsensitiveSet, format_url
+
 
 class UserType(Enum):
     User = 0
@@ -15,34 +16,45 @@ class UserType(Enum):
     Admin = 3
     OperationsAdmin = 4
 
+    def __str__(self):
+        return self.name
+
+    @classmethod
+    def _missing_(cls, value: str):
+        for member in cls:
+            if member.name.lower() == value.replace(" ", "").lower():
+                return member
+        # default
+        raise ValueError("Invalid element type=" + value)
+
+
 class User(TM1Object):
     """ Abstraction of a TM1 User
     
     """
 
     def __init__(self, name: str, groups: Iterable[str], friendly_name: Optional[str] = None,
-                 password: Optional[str] = None, enabled: Optional[bool] = True, user_type: Optional[str] = None):
+                 password: Optional[str] = None, user_type: Union[UserType, str] = None, enabled: bool = None):
         self._name = name
         self._groups = CaseAndSpaceInsensitiveSet(*groups)
         self._friendly_name = friendly_name
         self._password = password
         self._enabled = enabled
-
+        self._user_type = user_type
+        # determine user_type
         if user_type is None:
-            if 'Admin' in self._groups:
-                self._user_type = UserType.admin
-            elif 'SecurityAdmin' in self._groups:
-                self._user_type = UserType.securityAdmin
-            elif 'DataAdmin' in self._groups:
-                self._user_type = UserType.dataAdmin
-            elif 'OperationsAdmin' in self._groups:
-                self._user_type = UserType.OperationsAdmin
+            if str(UserType.Admin) in self._groups:
+                self.user_type = UserType.Admin
+            elif str(UserType.SecurityAdmin) in self._groups:
+                self.user_type = UserType.SecurityAdmin
+            elif str(UserType.DataAdmin) in self._groups:
+                self.user_type = UserType.DataAdmin
+            elif str(UserType.OperationsAdmin) in self._groups:
+                self.user_type = UserType.OperationsAdmin
             else:
-                self._user_type = UserType.User
+                self.user_type = UserType.User
         else:
-            self._user_type = UserType[user_type]
-
-
+            self.user_type = user_type
 
     @property
     def name(self) -> str:
@@ -71,8 +83,7 @@ class User(TM1Object):
 
     @property
     def enabled(self) -> bool:
-        if self._enabled:
-            return self._enabled
+        return self._enabled
 
     @name.setter
     def name(self, value: str):
@@ -87,12 +98,18 @@ class User(TM1Object):
         self._password = value
 
     @enabled.setter
-    def enabled(self, value: bool):
+    def enabled(self, value: Union[bool, None]):
         self._enabled = value
 
     @user_type.setter
-    def user_type(self, value:str):
-        self._user_type = UserType[value]
+    def user_type(self, value: Union[str, UserType]):
+        if not isinstance(value, str) and not isinstance(value, UserType):
+            raise ValueError("argument 'user_type' must be of type str or UserType")
+
+        self._user_type = UserType(value)
+        # update groups as well, since TM1 doesn't react to change in user_type property
+        if self._user_type is not UserType.User:
+            self.add_group(str(self._user_type))
 
     def add_group(self, group_name: str):
         self._groups.add(group_name)
@@ -119,9 +136,9 @@ class User(TM1Object):
         """
         return cls(name=user_as_dict['Name'],
                    friendly_name=user_as_dict['FriendlyName'],
-                   groups=[group["Name"] for group in user_as_dict['Groups']],
-                   enabled= user_as_dict['Enabled'],
-                   user_type=user_as_dict['Type'])
+                   enabled=user_as_dict["Enabled"],
+                   user_type=user_as_dict["Type"],
+                   groups=[group["Name"] for group in user_as_dict['Groups']])
 
     @property
     def body(self) -> str:
@@ -135,11 +152,11 @@ class User(TM1Object):
         body_as_dict = collections.OrderedDict()
         body_as_dict['Name'] = self.name
         body_as_dict['FriendlyName'] = self.friendly_name or self.name
+        body_as_dict['Enabled'] = self._enabled
+        body_as_dict['Type'] = str(self._user_type)
         if self.password:
             body_as_dict['Password'] = self._password
         body_as_dict['Groups@odata.bind'] = [format_url("Groups('{}')", group)
                                              for group
                                              in self.groups]
-        body_as_dict['Enabled'] = self._enabled
-        body_as_dict['Type'] = self._user_type
         return json.dumps(body_as_dict, ensure_ascii=False)
