@@ -8,7 +8,7 @@ from requests import Response
 from TM1py.Objects.User import User
 from TM1py.Services.ObjectService import ObjectService
 from TM1py.Services.RestService import RestService
-from TM1py.Utils.Utils import format_url
+from TM1py.Utils.Utils import format_url, CaseAndSpaceInsensitiveSet, require_admin
 
 
 class SecurityService(ObjectService):
@@ -25,6 +25,7 @@ class SecurityService(ObjectService):
     def determine_actual_group_name(self, group_name: str, **kwargs) -> str:
         return self.determine_actual_object_name(object_class="Groups", object_name=group_name, **kwargs)
 
+    @require_admin
     def create_user(self, user: User, **kwargs) -> Response:
         """ Create a user on TM1 Server
 
@@ -34,6 +35,7 @@ class SecurityService(ObjectService):
         url = '/api/v1/Users'
         return self._rest.POST(url, user.body, **kwargs)
 
+    @require_admin
     def create_group(self, group_name: str, **kwargs) -> Response:
         """ Create a Security group in the TM1 Server
 
@@ -50,7 +52,9 @@ class SecurityService(ObjectService):
         :return: instance of TM1py.User
         """
         user_name = self.determine_actual_user_name(user_name, **kwargs)
-        url = format_url("/api/v1/Users('{}')?$expand=Groups", user_name)
+        url = format_url(
+            "/api/v1/Users('{}')?$select=Name,FriendlyName,Password,Type,Enabled&$expand=Groups",
+            user_name)
         response = self._rest.GET(url, **kwargs)
         return User.from_dict(response.json())
 
@@ -59,10 +63,11 @@ class SecurityService(ObjectService):
 
         :return: instance of TM1py.User
         """
-        url = "/api/v1/ActiveUser?$expand=Groups"
+        url = "/api/v1/ActiveUser?$select=Name,FriendlyName,Password,Type,Enabled&$expand=Groups"
         response = self._rest.GET(url, **kwargs)
         return User.from_dict(response.json())
 
+    @require_admin
     def update_user(self, user: User, **kwargs) -> Response:
         """ Update user on TM1 Server
 
@@ -76,6 +81,12 @@ class SecurityService(ObjectService):
         url = format_url("/api/v1/Users('{}')", user.name)
         return self._rest.PATCH(url, user.body, **kwargs)
 
+    def update_user_password(self, user_name: str, password: str, **kwargs) -> Response:
+        url = format_url("/api/v1/Users('{}')", user_name)
+        body = {"Password": password}
+        return self._rest.PATCH(url, json.dumps(body), **kwargs)
+
+    @require_admin
     def delete_user(self, user_name: str, **kwargs) -> Response:
         """ Delete user on TM1 Server
 
@@ -86,6 +97,7 @@ class SecurityService(ObjectService):
         url = format_url("/api/v1/Users('{}')", user_name)
         return self._rest.DELETE(url, **kwargs)
 
+    @require_admin
     def delete_group(self, group_name: str, **kwargs) -> Response:
         """ Delete a group in the TM1 Server
 
@@ -101,7 +113,7 @@ class SecurityService(ObjectService):
 
         :return: List of TM1py.User instances
         """
-        url = '/api/v1/Users?$expand=Groups'
+        url = '/api/v1/Users?$select=Name,FriendlyName,Password,Type,Enabled&$expand=Groups'
         response = self._rest.GET(url, **kwargs)
         users = [User.from_dict(user) for user in response.json()['value']]
         return users
@@ -122,7 +134,9 @@ class SecurityService(ObjectService):
         :param group_name:
         :return: List of TM1py.User instances
         """
-        url = format_url("/api/v1/Groups('{}')?$expand=Users($expand=Groups)", group_name)
+        url = format_url(
+            "/api/v1/Groups('{}')?$expand=Users($select=Name,FriendlyName,Password,Type,Enabled;$expand=Groups)",
+            group_name)
         response = self._rest.GET(url, **kwargs)
         users = [User.from_dict(user) for user in response.json()['Users']]
         return users
@@ -149,6 +163,7 @@ class SecurityService(ObjectService):
         response = self._rest.GET(url, **kwargs)
         return [group['Name'] for group in response.json()['value']]
 
+    @require_admin
     def add_user_to_groups(self, user_name: str, groups: Iterable[str], **kwargs) -> Response:
         """
         
@@ -167,6 +182,7 @@ class SecurityService(ObjectService):
         }
         return self._rest.PATCH(url, json.dumps(body), **kwargs)
 
+    @require_admin
     def remove_user_from_group(self, group_name: str, user_name: str, **kwargs) -> Response:
         """ Remove user from group in TM1 Server
 
@@ -189,8 +205,52 @@ class SecurityService(ObjectService):
         groups = [entry['Name'] for entry in response.json()['value']]
         return groups
 
+    @require_admin
     def security_refresh(self, **kwargs) -> Response:
         from TM1py.Services import ProcessService
         ti = "SecurityRefresh;"
         process_service = ProcessService(self._rest)
         return process_service.execute_ti_code(ti, **kwargs)
+
+    def user_exists(self, user_name: str, **kwargs) -> bool:
+        url = format_url("/api/v1/Users('{}')", user_name)
+        return self._exists(url, **kwargs)
+
+    def group_exists(self, group_name: str, **kwargs) -> bool:
+        url = format_url("/api/v1/Groups('{}')", group_name)
+        return self._exists(url, **kwargs)
+
+    def get_custom_security_groups(self, **kwargs) -> List[str]:
+        custom_groups = CaseAndSpaceInsensitiveSet(*self.get_all_groups(**kwargs))
+        custom_groups.discard('Admin')
+        custom_groups.discard('DataAdmin')
+        custom_groups.discard('SecurityAdmin')
+        custom_groups.discard('OperationsAdmin')
+        custom_groups.discard('}tp_Everyone')
+
+        return list(custom_groups)
+
+    def get_read_only_users(self, **kwargs) -> List[str]:
+        read_only_users = list()
+
+        mdx = """
+        SELECT
+        {[}ClientProperties].[ReadOnlyUser]} ON COLUMNS,
+        NON EMPTY {[}Clients].MEMBERS} ON ROWS
+        FROM [}ClientProperties]
+        """
+
+        from TM1py import CellService
+        cell_service = CellService(self._rest)
+
+        users_with_flag = cell_service.execute_mdx_rows_and_values(
+            mdx=mdx,
+            element_unique_names=False,
+            **kwargs)
+
+        for row, values in users_with_flag.items():
+            user = row[0]
+            read_only = values[0]
+            if read_only:
+                read_only_users.append(user)
+        return read_only_users
