@@ -2,6 +2,7 @@
 
 import functools
 import json
+from collections.abc import Iterable
 from datetime import datetime
 from typing import Dict, Optional
 
@@ -11,6 +12,7 @@ from requests import Response
 from TM1py.Services.ObjectService import ObjectService
 from TM1py.Services.RestService import RestService
 from TM1py.Utils import format_url
+from TM1py.Utils.Utils import CaseAndSpaceInsensitiveDict, require_admin
 
 
 def odata_track_changes_header(func):
@@ -73,29 +75,94 @@ class ServerService(ObjectService):
         self.mlog_last_delta_request = response.text[response.text.rfind("MessageLogEntries/!delta('"):-2]
         return response.json()['value']
 
-    def get_message_log_entries(self, reverse: bool = True, top: int = None, **kwargs) -> Dict:
-        reverse = 'true' if reverse else 'false'
-        url = '/api/v1/MessageLog(Reverse={})'.format(reverse)
+    @require_admin
+    def get_message_log_entries(self, reverse: bool = True, since: datetime = None,
+                                until: datetime = None, top: int = None, logger: str = None,
+                                level: str = None, msg_contains: Iterable = None, msg_contains_operator: str = 'and',
+                                **kwargs) -> Dict:
+        """
+        :param reverse: Boolean
+        :param since: of type datetime. If it doesn't have tz information, UTC is assumed.
+        :param until: of type datetime. If it doesn't have tz information, UTC is assumed.
+        :param top: Integer
+        :param logger: string, eg TM1.Server, TM1.Chore, TM1.Mdx.Interface, TM1.Process
+        :param level: string, ERROR, WARNING, INFO, DEBUG, UNKNOWN
+        :param msg_contains: iterable, find substring in log message; list of substrings will be queried as AND statement
+        :param msg_contains_operator: 'and' or 'or'
+
+        :param kwargs:
+        :return: Dict of server log
+        """
+        msg_contains_operator = msg_contains_operator.strip().lower()
+        if msg_contains_operator not in ("and", "or"):
+            raise ValueError("'msg_contains_operator' must be either 'AND' or 'OR'")
+
+        reverse = 'desc' if reverse else 'asc'
+        url = '/api/v1/MessageLogEntries?$orderby=TimeStamp {}'.format(reverse)
+
+        if since or until or logger or level or msg_contains:
+            log_filters = []
+
+            if since:
+                # If since doesn't have tz information, UTC is assumed
+                if not since.tzinfo:
+                    since = self.utc_localize_time(since)
+                log_filters.append(format_url("TimeStamp ge {}", since.strftime("%Y-%m-%dT%H:%M:%SZ")))
+
+            if until:
+                # If until doesn't have tz information, UTC is assumed
+                if not until.tzinfo:
+                    until = self.utc_localize_time(until)
+                log_filters.append(format_url("TimeStamp le {}", until.strftime("%Y-%m-%dT%H:%M:%SZ")))
+
+            if logger:
+                log_filters.append(format_url("Logger eq '{}'", logger))
+
+            if level:
+                level_dict = CaseAndSpaceInsensitiveDict(
+                    {'ERROR': 1, 'WARNING': 2, 'INFO': 3, 'DEBUG': 4, 'UNKNOWN': 5})
+                level_index = level_dict.get(level)
+                if level_index:
+                    log_filters.append("Level eq {}".format(level_index))
+
+            if msg_contains:
+                if isinstance(msg_contains, str):
+                    log_filters.append(format_url("contains(toupper(Message),toupper('{}'))", msg_contains))
+                else:
+                    msg_filters = [format_url("contains(toupper(Message),toupper('{}'))", wildcard)
+                                   for wildcard in msg_contains]
+                    log_filters.append("({})".format(f" {msg_contains_operator} ".join(msg_filters)))
+
+            url += "&$filter={}".format(" and ".join(log_filters))
+
         if top:
-            url += '?$top={}'.format(top)
+            url += '&$top={}'.format(top)
+
         response = self._rest.GET(url, **kwargs)
         return response.json()['value']
 
+    @staticmethod
+    def utc_localize_time(timestamp):
+        timestamp = pytz.utc.localize(timestamp)
+        timestamp_utc = timestamp.astimezone(pytz.utc)
+        return timestamp_utc
+
+    @require_admin
     def get_transaction_log_entries(self, reverse: bool = True, user: str = None, cube: str = None,
-                                    since: datetime = None, top: int = None, **kwargs) -> Dict:
+                                    since: datetime = None, until: datetime = None, top: int = None, **kwargs) -> Dict:
         """
-        
-        :param reverse: 
-        :param user: 
-        :param cube: 
+        :param reverse: Boolean
+        :param user: UserName
+        :param cube: CubeName
         :param since: of type datetime. If it doesn't have tz information, UTC is assumed.
-        :param top: 
-        :return: 
+        :param until: of type datetime. If it doesn't have tz information, UTC is assumed.
+        :param top: int
+        :return:
         """
         reverse = 'desc' if reverse else 'asc'
         url = '/api/v1/TransactionLogEntries?$orderby=TimeStamp {} '.format(reverse)
         # filter on user, cube and time
-        if user or cube or since:
+        if user or cube or since or until:
             log_filters = []
             if user:
                 log_filters.append(format_url("User eq '{}'", user))
@@ -104,10 +171,13 @@ class ServerService(ObjectService):
             if since:
                 # If since doesn't have tz information, UTC is assumed
                 if not since.tzinfo:
-                    since = pytz.utc.localize(since)
-                # TM1 REST API expects %Y-%m-%dT%H:%M:%SZ Format with UTC time !
-                since_utc = since.astimezone(pytz.utc)
-                log_filters.append(format_url("TimeStamp ge {}", since_utc.strftime("%Y-%m-%dT%H:%M:%SZ")))
+                    since = self.utc_localize_time(since)
+                log_filters.append(format_url("TimeStamp ge {}", since.strftime("%Y-%m-%dT%H:%M:%SZ")))
+            if until:
+                # If until doesn't have tz information, UTC is assumed
+                if not until.tzinfo:
+                    until = self.utc_localize_time(until)
+                log_filters.append(format_url("TimeStamp le {}", until.strftime("%Y-%m-%dT%H:%M:%SZ")))
             url += "&$filter={}".format(" and ".join(log_filters))
         # top limit
         if top:
@@ -115,6 +185,7 @@ class ServerService(ObjectService):
         response = self._rest.GET(url, **kwargs)
         return response.json()['value']
 
+    @require_admin
     def get_last_process_message_from_messagelog(self, process_name: str, **kwargs) -> Optional[str]:
         """ Get the latest message log entry for a process
 
@@ -162,6 +233,7 @@ class ServerService(ObjectService):
         del config["@odata.context"]
         return config
 
+    @require_admin
     def get_static_configuration(self, **kwargs) -> Dict:
         """ Read TM1 config settings as dictionary from TM1 Server
 
@@ -172,6 +244,7 @@ class ServerService(ObjectService):
         del config["@odata.context"]
         return config
 
+    @require_admin
     def get_active_configuration(self, **kwargs) -> Dict:
         """ Read effective(!) TM1 config settings as dictionary from TM1 Server
 
@@ -182,6 +255,7 @@ class ServerService(ObjectService):
         del config["@odata.context"]
         return config
 
+    @require_admin
     def update_static_configuration(self, configuration: Dict) -> Response:
         """ Update the .cfg file and triggers TM1 to re-read the file.
 
@@ -191,8 +265,23 @@ class ServerService(ObjectService):
         url = '/api/v1/StaticConfiguration'
         return self._rest.PATCH(url, json.dumps(configuration))
 
+    @require_admin
     def save_data(self, **kwargs) -> Response:
         from TM1py.Services import ProcessService
         ti = "SaveDataAll;"
         process_service = ProcessService(self._rest)
         return process_service.execute_ti_code(ti, **kwargs)
+
+    @require_admin
+    def start_performance_monitor(self):
+        config = {
+            "Administration": {"PerformanceMonitorOn": True}
+        }
+        self.update_static_configuration(config)
+
+    @require_admin
+    def stop_performance_monitor(self):
+        config = {
+            "Administration": {"PerformanceMonitorOn": False}
+        }
+        self.update_static_configuration(config)
