@@ -1,6 +1,8 @@
 import pickle
 
-from TM1py.Services import *
+from TM1py.Services import HierarchyService, SecurityService, ApplicationService, SubsetService, ServerService, \
+    MonitoringService, ProcessService, PowerBiService, AnnotationService, ViewService, RestService, CellService, \
+    ChoreService, DimensionService, CubeService, ElementService, SandboxService
 
 
 class TM1Service:
@@ -25,7 +27,10 @@ class TM1Service:
         self.processes = ProcessService(self._tm1_rest)
         self.security = SecurityService(self._tm1_rest)
         self.server = ServerService(self._tm1_rest)
+        self.subsets = SubsetService(self._tm1_rest)
         self.applications = ApplicationService(self._tm1_rest)
+        self.views = ViewService(self._tm1_rest)
+        self.sandboxes = SandboxService(self._tm1_rest)
 
     def logout(self, **kwargs):
         self._tm1_rest.logout(**kwargs)
