@@ -1,3 +1,2 @@
-from TM1py.Utils.Utils import *
 from TM1py.Utils.MDXUtils import *
-from TM1py.Utils.TIObfuscator import *
+from TM1py.Utils.Utils import *
