@@ -6,21 +6,24 @@ from TM1py.Objects import Dimension, Hierarchy, Element
 from TM1py.Objects import ElementAttribute
 from TM1py.Services import TM1Service
 
-config = configparser.ConfigParser()
-config.read(Path(__file__).parent.joinpath('config.ini'))
 
-PREFIX = "TM1py_Tests_Dimension_"
-DIMENSION_NAME = PREFIX + "Some_Dimension"
-HIERARCHY_NAME = DIMENSION_NAME
-DIMENSION_NAME_WITH_MULTI_HIERARCHY = PREFIX + "Dimension_With_Multiple_Hierarchies"
-
-
-class TestDimensionMethods(unittest.TestCase):
-    tm1 = None
+class TestDimensionService(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.tm1 = TM1Service(**config['tm1srv01'])
+        """
+        Establishes a connection to TM1 and creates objects to use across all tests
+        """
+
+        # Connection to TM1
+        cls.config = configparser.ConfigParser()
+        cls.config.read(Path(__file__).parent.joinpath('config.ini'))
+        cls.tm1 = TM1Service(**cls.config['tm1srv01'])
+
+        cls.prefix = "TM1py_Tests_Dimension_"
+        cls.dimension_name = cls.prefix + "Some_Dimension"
+        cls.hierarchy_name = cls.dimension_name
+        cls.dimension_name_with_multi_hierarchy = cls.prefix + "Dimension_With_Multiple_Hierarchies"
 
     @classmethod
     def setUp(cls):
@@ -43,17 +46,17 @@ class TestDimensionMethods(unittest.TestCase):
             ElementAttribute(name='Name Long', attribute_type='Alias'),
             ElementAttribute(name='Name Short', attribute_type='Alias')]
         h = Hierarchy(
-            name=DIMENSION_NAME,
-            dimension_name=DIMENSION_NAME,
+            name=cls.dimension_name,
+            dimension_name=cls.dimension_name,
             elements=elements,
             edges=edges,
             element_attributes=element_attributes)
-        d = Dimension(name=DIMENSION_NAME, hierarchies=[h])
+        d = Dimension(name=cls.dimension_name, hierarchies=[h])
         cls.tm1.dimensions.create(d)
 
     @classmethod
     def create_dimension_with_multiple_hierarchies(cls):
-        dimension = Dimension(DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = Dimension(cls.dimension_name_with_multi_hierarchy)
         dimension.add_hierarchy(
             Hierarchy(
                 name="Hierarchy1",
@@ -73,47 +76,47 @@ class TestDimensionMethods(unittest.TestCase):
 
     @classmethod
     def delete_dimensions(cls):
-        cls.tm1.dimensions.delete(DIMENSION_NAME)
-        if cls.tm1.dimensions.exists(DIMENSION_NAME_WITH_MULTI_HIERARCHY):
-            cls.tm1.dimensions.delete(DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        cls.tm1.dimensions.delete(cls.dimension_name)
+        if cls.tm1.dimensions.exists(cls.dimension_name_with_multi_hierarchy):
+            cls.tm1.dimensions.delete(cls.dimension_name_with_multi_hierarchy)
 
     def test_get_dimension(self):
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
         self.assertIsInstance(d, Dimension)
-        self.assertEqual(d.name, DIMENSION_NAME)
+        self.assertEqual(d.name, self.dimension_name)
         h = d.hierarchies[0]
         self.assertIsInstance(h, Hierarchy)
-        self.assertEqual(h.name, DIMENSION_NAME)
+        self.assertEqual(h.name, self.dimension_name)
         self.assertEqual(len(h.elements), 1001)
         self.assertEqual(len(h.edges), 1000)
 
     def test_dimension__get__(self):
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
-        h = d[DIMENSION_NAME]
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
+        h = d[self.dimension_name]
         self.assertIsInstance(h, Hierarchy)
-        self.assertEqual(h.name, DIMENSION_NAME)
+        self.assertEqual(h.name, self.dimension_name)
         self.assertEqual(len(h.elements), 1001)
         self.assertEqual(len(h.edges), 1000)
 
     def test_dimension__contains__(self):
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
-        self.assertIn(DIMENSION_NAME, d)
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
+        self.assertIn(self.dimension_name, d)
 
     def test_dimension__iter__(self):
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
         first_hierarchy = next(h for h in d)
         self.assertIsInstance(first_hierarchy, Hierarchy)
-        self.assertEqual(first_hierarchy.name, DIMENSION_NAME)
+        self.assertEqual(first_hierarchy.name, self.dimension_name)
         self.assertEqual(len(first_hierarchy.elements), 1001)
         self.assertEqual(len(first_hierarchy.edges), 1000)
 
     def test_dimension__len__(self):
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
         self.assertEqual(len(d), 1)
 
     def test_update_dimension(self):
         # get dimension from tm1
-        d = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME)
+        d = self.tm1.dimensions.get(dimension_name=self.dimension_name)
         # create element objects
         elements = [Element(name='e1', element_type='Consolidated'),
                     Element(name='e2', element_type='Numeric'),
@@ -129,49 +132,53 @@ class TestDimensionMethods(unittest.TestCase):
                               ElementAttribute(name='Name Short', attribute_type='Alias'),
                               ElementAttribute(name='Currency', attribute_type='String')]
         # create hierarchy object
-        hierarchy = Hierarchy(name=DIMENSION_NAME, dimension_name=DIMENSION_NAME, elements=elements,
+        hierarchy = Hierarchy(name=self.dimension_name, dimension_name=self.dimension_name, elements=elements,
                               element_attributes=element_attributes, edges=edges)
 
         # replace existing hierarchy with new hierarchy
-        d.remove_hierarchy(DIMENSION_NAME)
+        d.remove_hierarchy(self.dimension_name)
         d.add_hierarchy(hierarchy)
 
         # update dimension in TM1
         self.tm1.dimensions.update(d)
 
         # Test
-        dimension = self.tm1.dimensions.get(DIMENSION_NAME)
+        dimension = self.tm1.dimensions.get(self.dimension_name)
         self.assertEqual(len(dimension.hierarchies[0].elements), len(elements))
 
     def test_update_dimension_remove_hierarchy(self):
         self.create_dimension_with_multiple_hierarchies()
-        dimension = self.tm1.dimensions.get(DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(self.dimension_name_with_multi_hierarchy)
         self.assertEqual(dimension.hierarchy_names, ['Hierarchy1', 'Hierarchy2', 'Hierarchy3', 'Leaves'])
         dimension.remove_hierarchy('Hierarchy2')
         dimension.remove_hierarchy('Hierarchy3')
         self.tm1.dimensions.update(dimension)
-        dimension = self.tm1.dimensions.get(DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(self.dimension_name_with_multi_hierarchy)
         self.assertEqual(dimension.hierarchy_names, ['Hierarchy1', 'Leaves'])
 
     def test_get_all_names(self):
-        self.assertIn(DIMENSION_NAME, self.tm1.dimensions.get_all_names())
+        self.assertIn(self.dimension_name, self.tm1.dimensions.get_all_names())
+
+    def test_get_number_of_dimensions(self):
+        number_of_dimensions = self.tm1.dimensions.get_number_of_dimensions()
+        self.assertIsInstance(number_of_dimensions, int)
 
     def test_execute_mdx(self):
-        mdx = "{TM1SubsetAll(" + DIMENSION_NAME + ")}"
-        elements = self.tm1.dimensions.execute_mdx(DIMENSION_NAME, mdx)
+        mdx = "{TM1SubsetAll(" + self.dimension_name + ")}"
+        elements = self.tm1.dimensions.execute_mdx(self.dimension_name, mdx)
         self.assertEqual(len(elements), 1001)
 
-        mdx = "{ Tm1FilterByLevel ( {TM1SubsetAll(" + DIMENSION_NAME + ")}, 0) }"
-        elements = self.tm1.dimensions.execute_mdx(DIMENSION_NAME, mdx)
+        mdx = "{ Tm1FilterByLevel ( {TM1SubsetAll(" + self.dimension_name + ")}, 0) }"
+        elements = self.tm1.dimensions.execute_mdx(self.dimension_name, mdx)
         self.assertEqual(len(elements), 1000)
         for element in elements:
-            element.startswith("Element")
+            self.assertTrue(element.startswith("Element"))
 
     def test_hierarchy_names(self):
         # create dimension with two Hierarchies
         self.create_dimension_with_multiple_hierarchies()
 
-        dimension = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(dimension_name=self.dimension_name_with_multi_hierarchy)
         self.assertEqual(
             set(dimension.hierarchy_names),
             {"Leaves", "Hierarchy1", "Hierarchy2", "Hierarchy3"})
@@ -184,7 +191,7 @@ class TestDimensionMethods(unittest.TestCase):
     def test_remove_leaves_hierarchy(self):
         # create dimension with two Hierarchies
         self.create_dimension_with_multiple_hierarchies()
-        dimension = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(dimension_name=self.dimension_name_with_multi_hierarchy)
 
         try:
             dimension.remove_hierarchy("LEAVES")
@@ -196,7 +203,7 @@ class TestDimensionMethods(unittest.TestCase):
         # create dimension with two Hierarchies
         self.create_dimension_with_multiple_hierarchies()
 
-        dimension = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(dimension_name=self.dimension_name_with_multi_hierarchy)
         self.assertEqual(len(dimension.hierarchies), 4)
         self.assertIn("Hierarchy1", dimension)
         self.assertIn("Hierarchy2", dimension)
@@ -206,7 +213,7 @@ class TestDimensionMethods(unittest.TestCase):
         dimension.remove_hierarchy("Hierarchy1")
         self.tm1.dimensions.update(dimension)
 
-        dimension = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(dimension_name=self.dimension_name_with_multi_hierarchy)
         self.assertEqual(len(dimension.hierarchies), 3)
         self.assertNotIn("Hierarchy1", dimension)
         self.assertIn("Hierarchy2", dimension)
@@ -216,7 +223,7 @@ class TestDimensionMethods(unittest.TestCase):
         dimension.remove_hierarchy("H i e r a r c h y 3".upper())
         self.tm1.dimensions.update(dimension)
 
-        dimension = self.tm1.dimensions.get(dimension_name=DIMENSION_NAME_WITH_MULTI_HIERARCHY)
+        dimension = self.tm1.dimensions.get(dimension_name=self.dimension_name_with_multi_hierarchy)
         self.assertEqual(len(dimension.hierarchies), 2)
         self.assertNotIn("Hierarchy1", dimension)
         self.assertIn("Hierarchy2", dimension)
@@ -224,8 +231,8 @@ class TestDimensionMethods(unittest.TestCase):
         self.assertIn("Leaves", dimension)
 
     def test_rename_dimension(self):
-        original_dimension_name = PREFIX + "Original_Dimension"
-        renamed_dimension_name = PREFIX + "Renamed_Dimension"
+        original_dimension_name = self.prefix + "Original_Dimension"
+        renamed_dimension_name = self.prefix + "Renamed_Dimension"
 
         # if dimensions exist in TM1.. delete them
         for dim_name in (original_dimension_name, renamed_dimension_name):
