@@ -5,13 +5,23 @@ import unittest
 from TM1py.Services import TM1Service
 from TM1py.Utils.Utils import CaseAndSpaceInsensitiveSet, CaseAndSpaceInsensitiveDict, CaseAndSpaceInsensitiveTuplesDict
 
-config = configparser.ConfigParser()
-config.read(Path(__file__).parent.joinpath('config.ini'))
-
 
 class TestCaseAndSpaceInsensitiveDict(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        """
+        Establishes a connection to TM1 and creates TM! objects to use across all tests
+        """
+
+        # Connection to TM1
+        cls.config = configparser.ConfigParser()
+        cls.config.read(Path(__file__).parent.joinpath('config.ini'))
+        cls.tm1 = TM1Service(**cls.config['tm1srv01'])
+        
     @classmethod
     def setUp(cls):
+
         cls.map = CaseAndSpaceInsensitiveDict()
         cls.map["key1"] = "value1"
         cls.map["key2"] = "value2"
@@ -204,6 +214,13 @@ class TestCaseAndSpaceInsensitiveTuplesDict(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
+        """
+        Establishes a connection to TM1 and creates TM! objects to use across all tests
+        """
+
+        # Connection to TM1
+        config = configparser.ConfigParser()
+        config.read(Path(__file__).parent.joinpath('config.ini'))
         cls.tm1 = TM1Service(**config['tm1srv01'])
 
     def setUp(self):
